refactor(rib): split manager/distribution.rs into concern submodules
