+++ ROADMAP.md
@@ -1393,7 +1393,9 @@ branch is between features.
   daemon bin out of the doc target set (avoiding the lib/bin same-name collision);
   Cargo's default job parallelism is intentionally left enabled so rustdoc does
   not serialize the whole workspace.
-- [ ] **Mega-module splits.** The large `src/` modules have been split, but
+- [ ] **Mega-module splits.** The large `src/` modules have been split, and
+  `crates/rib/src/manager/distribution.rs` (which had absorbed the BGP-LS/VPN/
+  RTC/ORR arcs) is now a directory module with per-family concern submodules.
   `crates/api/src/event_service.rs` remains borderline. Keep splitting only where
   it reduces real conflict or review cost.
 - [ ] **Doc-precision + lint-policy consistency sweep (v0.41.0 review).** A
