--- crates/rib/src/manager/distribution.rs
@@ -1,3582 +1,0 @@
-use std::collections::{HashMap, HashSet};
-use std::net::{IpAddr, Ipv4Addr};
-use std::sync::Arc;
-
-use rustbgpd_policy::{
-    PolicyAction, PolicyChain, PolicyEvaluation, RouteContext, RouteType,
-    evaluate_chain_with_attribution,
-};
-use rustbgpd_rpki::VrpTable;
-use rustbgpd_telemetry::BgpMetrics;
-use rustbgpd_wire::{
-    AddressPrefixOrf, Afi, FlowSpecRule, OrfAction, OrfMatch, Prefix, RouteRefreshSubtype, Safi,
-    WhenToRefresh,
-};
-use tracing::{debug, info, warn};
-
-use super::helpers::{
-    LOCAL_PEER, bgpls_routes_equal, evpn_routes_equal, gauge_val, prefix_family, routes_equal,
-    rtc_routes_equal, should_suppress_ibgp_inner, validate_route_aspa, validate_route_rpki,
-    vpn_routes_equal,
-};
-use super::{PendingRouteChunk, PendingRoutesReceived, PolicyFilteredRouteKey, RibManager};
-use crate::adj_rib_in::AdjRibIn;
-use crate::adj_rib_out::AdjRibOut;
-use crate::event::{RouteEvent, RouteEventType};
-use crate::loc_rib::LocRib;
-use crate::update::{
-    ExplainAdvertisedRoute, ExplainDecision, ExplainReason, NeighborPolicyStats,
-    OutboundRouteUpdate, RibCommandError,
-};
-
-fn route_type(origin: crate::route::RouteOrigin) -> RouteType {
-    match origin {
-        crate::route::RouteOrigin::Local => RouteType::Local,
-        crate::route::RouteOrigin::Ibgp => RouteType::Internal,
-        crate::route::RouteOrigin::Ebgp => RouteType::External,
-    }
-}
-
-fn route_type_label(route_type: RouteType) -> &'static str {
-    match route_type {
-        RouteType::Local => "local_route",
-        RouteType::Internal => "ibgp_route",
-        RouteType::External => "ebgp_route",
-    }
-}
-
-fn route_type_message(route_type: RouteType) -> &'static str {
-    match route_type {
-        RouteType::Local => "best route is locally originated",
-        RouteType::Internal => "best route was learned from an iBGP peer",
-        RouteType::External => "best route was learned from an eBGP peer",
-    }
-}
-
-fn record_export_policy_eval(
-    metrics: &BgpMetrics,
-    stats: &mut NeighborPolicyStats,
-    peer_label: &str,
-    evaluation: &PolicyEvaluation,
-) {
-    let policy = evaluation.matched_policy.as_deref().unwrap_or("inline");
-    let action = match evaluation.action {
-        PolicyAction::Permit => {
-            stats.export_policy_routes_permitted =
-                stats.export_policy_routes_permitted.saturating_add(1);
-            "permit"
-        }
-        PolicyAction::Deny => {
-            stats.export_policy_routes_denied = stats.export_policy_routes_denied.saturating_add(1);
-            "deny"
-        }
-    };
-    metrics.record_policy_routes(peer_label, policy, "export", action);
-}
-
-impl RibManager {
-    #[expect(
-        clippy::too_many_arguments,
-        clippy::too_many_lines,
-        reason = "explain mirrors live single-best export inputs for policy parity"
-    )]
-    pub(super) fn explain_single_best_prefix(
-        loc_rib: &LocRib,
-        peer_is_rr_client: &HashMap<IpAddr, bool>,
-        prefix: Prefix,
-        target_peer: IpAddr,
-        target_peer_asn: Option<u32>,
-        target_peer_group: Option<&str>,
-        target_is_ebgp: bool,
-        target_is_rr_client: bool,
-        cluster_id: Option<Ipv4Addr>,
-        sendable: Option<&Vec<(Afi, Safi)>>,
-        export_pol: Option<&PolicyChain>,
-    ) -> ExplainAdvertisedRoute {
-        let mut explain = ExplainAdvertisedRoute {
-            decision: ExplainDecision::NoBestRoute,
-            peer: target_peer,
-            prefix,
-            next_hop: None,
-            path_id: 0,
-            route_peer: None,
-            route_type: None,
-            reasons: Vec::new(),
-            modifications: rustbgpd_policy::RouteModifications::default(),
-        };
-
-        let Some(best) = loc_rib.get(&prefix) else {
-            explain.reasons.push(ExplainReason {
-                code: "no_best_route",
-                message: "no best route exists for this prefix".to_string(),
-            });
-            return explain;
-        };
-
-        explain.route_peer = Some(best.peer);
-        explain.route_type = Some(route_type(best.origin_type));
-
-        if best.peer == target_peer {
-            explain.decision = ExplainDecision::Deny;
-            explain.reasons.push(ExplainReason {
-                code: "ibgp_split_horizon",
-                message: "route is suppressed by split horizon because it originated from the target peer".to_string(),
-            });
-            return explain;
-        }
-
-        if should_suppress_ibgp_inner(
-            best,
-            target_is_ebgp,
-            target_is_rr_client,
-            cluster_id,
-            peer_is_rr_client,
-        ) {
-            explain.decision = ExplainDecision::Deny;
-            let code = if !target_is_ebgp
-                && !best.is_ebgp()
-                && cluster_id.is_some()
-                && !peer_is_rr_client.get(&best.peer).copied().unwrap_or(false)
-                && !target_is_rr_client
-            {
-                "rr_non_client_to_non_client"
-            } else {
-                "ibgp_split_horizon"
-            };
-            let message = if code == "rr_non_client_to_non_client" {
-                "route reflector will not reflect a non-client iBGP route to another non-client"
-                    .to_string()
-            } else {
-                "iBGP split horizon suppresses advertisement of this route".to_string()
-            };
-            explain.reasons.push(ExplainReason { code, message });
-            return explain;
-        }
-
-        let family = prefix_family(&prefix);
-        if !sendable.is_some_and(|f| f.contains(&family)) {
-            explain.decision = ExplainDecision::UnsupportedFamily;
-            explain.reasons.push(ExplainReason {
-                code: "family_not_sendable",
-                message: format!(
-                    "peer cannot receive {} {} routes",
-                    match family.0 {
-                        Afi::Ipv4 => "ipv4",
-                        Afi::Ipv6 => "ipv6",
-                        Afi::L2Vpn => "l2vpn",
-                        Afi::BgpLs => "bgpls",
-                    },
-                    match family.1 {
-                        Safi::Unicast => "unicast",
-                        Safi::FlowSpec => "flowspec",
-                        Safi::Multicast => "multicast",
-                        Safi::Evpn => "evpn",
-                        Safi::BgpLs => "bgpls",
-                        Safi::BgpLsVpn => "bgpls_vpn",
-                        Safi::MplsVpn => "mpls_vpn",
-                        Safi::RtConstrain => "rtc",
-                    }
-                ),
-            });
-            return explain;
-        }
-
-        let aspath_str = if export_pol.is_some_and(PolicyChain::requires_as_path_string) {
-            best.as_path()
-                .map_or_else(String::new, rustbgpd_wire::AsPath::to_aspath_string)
-        } else {
-            String::new()
-        };
-        let aspath_len = best.as_path().map_or(0, rustbgpd_wire::AsPath::len);
-        let ctx = RouteContext {
-            prefix: Some(prefix),
-            next_hop: Some(best.next_hop),
-            extended_communities: best.extended_communities(),
-            communities: best.communities(),
-            large_communities: best.large_communities(),
-            as_path_str: &aspath_str,
-            as_path_len: aspath_len,
-            validation_state: best.validation_state,
-            aspa_state: best.aspa_state,
-            peer_address: Some(target_peer),
-            peer_asn: target_peer_asn,
-            peer_group: target_peer_group,
-            route_type: explain.route_type,
-            evpn_route_type: None,
-            local_pref: best.local_pref_attr(),
-            med: best.med_attr(),
-        };
-        // Explain is a one-shot operator query path: enrich the deny /
-        // permit reason with the terminal-decision policy name but do
-        // NOT increment bgp_policy_routes_total here — the actual
-        // distribution path counts each route once, and double-counting
-        // explain calls would skew the metric.
-        let (result, evaluation) = evaluate_chain_with_attribution(export_pol, &ctx);
-        let policy_label = evaluation.matched_policy.as_deref().unwrap_or("inline");
-        if result.action != PolicyAction::Permit {
-            explain.decision = ExplainDecision::Deny;
-            explain.reasons.push(ExplainReason {
-                code: "policy_denied",
-                message: format!("export policy {policy_label:?} denied this route"),
-            });
-            return explain;
-        }
-
-        explain.decision = ExplainDecision::Advertise;
-        explain.modifications = result.modifications.clone();
-        if let Some(route_type) = explain.route_type {
-            explain.reasons.push(ExplainReason {
-                code: route_type_label(route_type),
-                message: route_type_message(route_type).to_string(),
-            });
-        }
-        if export_pol.is_some() {
-            explain.reasons.push(ExplainReason {
-                code: "policy_permitted",
-                message: format!("export policy {policy_label:?} permitted this route"),
-            });
-        }
-
-        let mut next_hop = best.next_hop;
-        if let Some(rustbgpd_policy::NextHopAction::Specific(addr)) =
-            explain.modifications.set_next_hop.clone()
-        {
-            next_hop = addr;
-        }
-        explain.next_hop = Some(next_hop);
-        explain
-    }
-
-    #[expect(
-        clippy::too_many_arguments,
-        clippy::too_many_lines,
-        reason = "outbound commit needs all family queues for one atomic send"
-    )]
-    pub(super) fn try_send_and_commit_outbound_update(
-        &mut self,
-        peer: IpAddr,
-        next_hop_override: Vec<Option<rustbgpd_policy::NextHopAction>>,
-        announce: Vec<crate::route::Route>,
-        withdraw: Vec<(Prefix, u32)>,
-        end_of_rib: Vec<(Afi, Safi)>,
-        refresh_markers: Vec<(Afi, Safi, RouteRefreshSubtype)>,
-        flowspec_announce: Vec<crate::route::FlowSpecRoute>,
-        flowspec_withdraw: Vec<FlowSpecRule>,
-        evpn_announce: Vec<crate::route::EvpnRibRoute>,
-        evpn_withdraw: Vec<rustbgpd_wire::EvpnRouteKey>,
-        bgpls_announce: Vec<crate::route::BgpLsRibRoute>,
-        bgpls_withdraw: Vec<crate::route::BgpLsRouteKey>,
-        vpn_announce: Vec<crate::route::VpnRibRoute>,
-        vpn_withdraw: Vec<crate::route::VpnRibRouteKey>,
-        rtc_announce: Vec<crate::route::RtcRibRoute>,
-        rtc_withdraw: Vec<crate::route::RtcRibRouteKey>,
-    ) -> bool {
-        let Some(tx) = self.outbound_peers.get(&peer).cloned() else {
-            return false;
-        };
-        let Ok(permit) = tx.try_reserve() else {
-            return false;
-        };
-
-        if !announce.is_empty()
-            || !withdraw.is_empty()
-            || !flowspec_announce.is_empty()
-            || !flowspec_withdraw.is_empty()
-            || !evpn_announce.is_empty()
-            || !evpn_withdraw.is_empty()
-            || !bgpls_announce.is_empty()
-            || !bgpls_withdraw.is_empty()
-            || !vpn_announce.is_empty()
-            || !vpn_withdraw.is_empty()
-            || !rtc_announce.is_empty()
-            || !rtc_withdraw.is_empty()
-        {
-            let loc_rib_len = self.loc_rib.len();
-            let rib_out = self
-                .adj_ribs_out
-                .entry(peer)
-                .or_insert_with(|| AdjRibOut::with_capacity(peer, loc_rib_len));
-            for route in &announce {
-                rib_out.insert(route.clone());
-            }
-            for (prefix, path_id) in &withdraw {
-                rib_out.withdraw(prefix, *path_id);
-            }
-            for route in &flowspec_announce {
-                rib_out.insert_flowspec(route.clone());
-            }
-            for rule in &flowspec_withdraw {
-                rib_out.remove_flowspec(rule);
-            }
-            for route in &evpn_announce {
-                rib_out.insert_evpn(route.clone());
-            }
-            for key in &evpn_withdraw {
-                rib_out.remove_evpn(key);
-            }
-            for route in &bgpls_announce {
-                rib_out.insert_bgpls(route.clone());
-            }
-            for key in &bgpls_withdraw {
-                rib_out.remove_bgpls(key);
-            }
-            for route in &vpn_announce {
-                rib_out.insert_vpn(route.clone());
-            }
-            for key in &vpn_withdraw {
-                rib_out.remove_vpn(key);
-            }
-            for route in &rtc_announce {
-                rib_out.insert_rtc(route.clone());
-            }
-            for key in &rtc_withdraw {
-                rib_out.remove_rtc(key);
-            }
-            self.metrics.set_adj_rib_out_prefixes(
-                &peer.to_string(),
-                "all",
-                gauge_val(rib_out.len()),
-            );
-            self.metrics.set_adj_rib_out_prefixes(
-                &peer.to_string(),
-                "flowspec",
-                gauge_val(rib_out.flowspec_len()),
-            );
-            self.metrics.set_adj_rib_out_prefixes(
-                &peer.to_string(),
-                "evpn",
-                gauge_val(rib_out.evpn_len()),
-            );
-            self.metrics.set_adj_rib_out_prefixes(
-                &peer.to_string(),
-                "bgpls",
-                gauge_val(rib_out.bgpls_len()),
-            );
-            self.metrics.set_adj_rib_out_prefixes(
-                &peer.to_string(),
-                "vpn",
-                gauge_val(rib_out.vpn_len()),
-            );
-            self.metrics.set_adj_rib_out_prefixes(
-                &peer.to_string(),
-                "rtc",
-                gauge_val(rib_out.rtc_len()),
-            );
-        }
-
-        permit.send(OutboundRouteUpdate {
-            announce,
-            withdraw,
-            end_of_rib,
-            refresh_markers,
-            next_hop_override,
-            flowspec_announce,
-            flowspec_withdraw,
-            evpn_announce,
-            evpn_withdraw,
-            bgpls_announce,
-            bgpls_withdraw,
-            vpn_announce,
-            vpn_withdraw,
-            rtc_announce,
-            rtc_withdraw,
-            request_refresh_all_negotiated: false,
-        });
-        true
-    }
-
-    pub(super) fn handle_replace_peer_export_policy(
-        &mut self,
-        peer: IpAddr,
-        export_policy: Option<PolicyChain>,
-        reply: tokio::sync::oneshot::Sender<Result<(), String>>,
-    ) {
-        if !self.outbound_peers.contains_key(&peer) {
-            let _ = reply.send(Err(format!(
-                "peer {peer} not registered for outbound updates"
-            )));
-            return;
-        }
-
-        self.peer_export_policies.insert(peer, export_policy);
-        self.dirty_peers.insert(peer);
-        self.distribute_changes(&HashSet::new(), &HashSet::new());
-        let _ = reply.send(Ok(()));
-    }
-
-    /// Force re-emission of all currently-advertised routes to a peer
-    /// without changing policy. The export-policy path already does
-    /// the same thing as a side-effect of policy replacement; this
-    /// variant exists for outbound *attribute* surface changes that
-    /// don't go through policy (e.g. RFC 8326 `GShut` community attach
-    /// toggle, where the toggle lives as a per-session bool but
-    /// changing it must trigger a fresh outbound emission so peers
-    /// see the updated wire form on routes already in `AdjRibOut`).
-    pub(super) fn handle_refresh_peer_outbound(
-        &mut self,
-        peer: IpAddr,
-        reply: tokio::sync::oneshot::Sender<Result<(), String>>,
-    ) {
-        if !self.outbound_peers.contains_key(&peer) {
-            let _ = reply.send(Err(format!(
-                "peer {peer} not registered for outbound updates"
-            )));
-            return;
-        }
-        // `force_outbound_peers` (not `dirty_peers`) — a force-only
-        // resync bypasses AdjRibOut equality suppression for currently-
-        // advertised exportable routes, which is exactly what GShut and
-        // similar outbound-attribute toggles need (the wire change is
-        // applied later in transport, invisible to this RIB diff). The
-        // distribution loop clears the entry on successful emission so
-        // the force is one-shot.
-        self.force_outbound_peers.insert(peer);
-        self.distribute_changes(&HashSet::new(), &HashSet::new());
-        let _ = reply.send(Ok(()));
-    }
-
-    /// Apply Address-Prefix ORF entries pushed by a peer (RFC 5291/5292).
-    ///
-    /// Installs/updates the per-`(peer, AFI, SAFI)` filter and lifts the
-    /// initial-advertisement gate. The re-advertisement sweep runs for
-    /// `IMMEDIATE` (or after a malformed-field reset, RFC 5291 §5.2); `DEFER`
-    /// installs the filter only — it stays live for subsequent outbound churn
-    /// and is swept on a later IMMEDIATE/plain ROUTE-REFRESH.
-    pub(super) fn handle_peer_orf_update(
-        &mut self,
-        peer: IpAddr,
-        afi: Afi,
-        safi: Safi,
-        when: rustbgpd_wire::WhenToRefresh,
-        entries: &[rustbgpd_wire::AddressPrefixOrf],
-        reply: tokio::sync::oneshot::Sender<Result<(), String>>,
-    ) {
-        if !self.outbound_peers.contains_key(&peer) {
-            let _ = reply.send(Err(format!(
-                "peer {peer} not registered for outbound updates"
-            )));
-            return;
-        }
-        let family = (afi, safi);
-        // The ORF message is itself a ROUTE-REFRESH (RFC 5291 §6) — lift the gate.
-        if let Some(pending) = self.peer_orf_pending.get_mut(&peer) {
-            pending.remove(&family);
-        }
-        let filter = self
-            .peer_orf_filters
-            .entry(peer)
-            .or_default()
-            .entry(family)
-            .or_default();
-        let unknown_when = matches!(when, WhenToRefresh::Unknown(_));
-        let reset = match when {
-            WhenToRefresh::Unknown(value) => {
-                // RFC 5291 only defines IMMEDIATE and DEFER. Treat an unknown
-                // value like a malformed negotiated ORF control field: clear the
-                // installed list and force a safe resync instead of silently
-                // installing peer state with defer-like behavior.
-                let remove_all = AddressPrefixOrf {
-                    action: OrfAction::RemoveAll,
-                    match_: OrfMatch::Permit,
-                    sequence: 0,
-                    min_len: 0,
-                    max_len: 0,
-                    prefix: None,
-                };
-                let _ = filter.apply(&[remove_all]);
-                warn!(
-                    %peer,
-                    ?family,
-                    when_to_refresh = value,
-                    "unknown ORF When-to-refresh — installed ORF list for this type reset"
-                );
-                true
-            }
-            _ => filter.apply(entries).is_err(),
-        };
-        let now_empty = filter.is_empty();
-        if reset && !unknown_when {
-            warn!(%peer, ?family, "malformed ORF entry — installed ORF list for this type reset");
-        }
-        // An emptied filter (REMOVE-ALL or a reset) means permit-all again —
-        // drop the entry so the absent-filter fast path applies.
-        if now_empty && let Some(by_family) = self.peer_orf_filters.get_mut(&peer) {
-            by_family.remove(&family);
-        }
-        if reset || when == WhenToRefresh::Immediate {
-            // If this peer's EoR for the family was withheld at `PeerUp`
-            // (GR restarter + §6 gate, see `send_initial_table`), the forced
-            // resync below is the gated flood: move the family into
-            // `pending_eor` and mark the peer dirty so the resync piggybacks
-            // the EoR behind the flooded routes (or flushes it standalone
-            // when the filter yields no routes). A DEFER update lifts the
-            // gate without flooding, so the deferral stays put and rides the
-            // later plain ROUTE-REFRESH or IMMEDIATE update that actually
-            // advertises the family.
-            if let Some(families) = self.gr_deferred_eor.get_mut(&peer)
-                && families.remove(&family)
-            {
-                if families.is_empty() {
-                    self.gr_deferred_eor.remove(&peer);
-                }
-                self.pending_eor.entry(peer).or_default().insert(family);
-                self.dirty_peers.insert(peer);
-            }
-            self.force_outbound_peers.insert(peer);
-            self.distribute_changes(&HashSet::new(), &HashSet::new());
-        }
-        let _ = reply.send(Ok(()));
-    }
-
-    #[expect(
-        clippy::too_many_arguments,
-        reason = "RIB update messages carry every supported family as one transaction"
-    )]
-    pub(super) fn enqueue_routes_received(
-        &mut self,
-        peer: IpAddr,
-        announced: Vec<crate::route::Route>,
-        withdrawn: Vec<(Prefix, u32)>,
-        flowspec_announced: Vec<crate::route::FlowSpecRoute>,
-        flowspec_withdrawn: Vec<FlowSpecRule>,
-        evpn_announced: Vec<crate::route::EvpnRibRoute>,
-        evpn_withdrawn: Vec<rustbgpd_wire::EvpnRouteKey>,
-    ) {
-        if let std::collections::hash_map::Entry::Vacant(entry) = self.ribs.entry(peer) {
-            let pending = PendingRoutesReceived::new(
-                peer,
-                announced,
-                withdrawn,
-                flowspec_announced,
-                flowspec_withdrawn,
-                evpn_announced,
-                evpn_withdrawn,
-            );
-            entry.insert(AdjRibIn::with_capacity(
-                peer,
-                pending.route_capacity_hint(),
-                pending.flowspec_capacity_hint(),
-            ));
-            self.pending_route_batches.push_back(pending);
-            return;
-        }
-
-        self.pending_route_batches
-            .push_back(PendingRoutesReceived::new(
-                peer,
-                announced,
-                withdrawn,
-                flowspec_announced,
-                flowspec_withdrawn,
-                evpn_announced,
-                evpn_withdrawn,
-            ));
-    }
-
-    pub(super) fn process_next_route_chunk(&mut self) -> bool {
-        let Some(mut pending) = self.pending_route_batches.pop_front() else {
-            return false;
-        };
-
-        let peer = pending.peer();
-        let Some(chunk) = pending.next_chunk() else {
-            // Empty/exhausted batch — flush anything still accumulated
-            // (defensive; normally the has_more() branch below flushes).
-            self.flush_pending_distribute();
-            return false;
-        };
-
-        match chunk {
-            PendingRouteChunk::Withdrawn(withdrawn) => {
-                self.process_withdraw_chunk(peer, withdrawn);
-            }
-            PendingRouteChunk::Announced(announced) => {
-                self.process_announce_chunk(peer, announced);
-            }
-            PendingRouteChunk::FlowSpecWithdrawn(flowspec_withdrawn) => {
-                self.process_flowspec_withdraw_chunk(peer, flowspec_withdrawn);
-            }
-            PendingRouteChunk::FlowSpecAnnounced(flowspec_announced) => {
-                self.process_flowspec_announce_chunk(peer, flowspec_announced);
-            }
-            PendingRouteChunk::EvpnWithdrawn(evpn_withdrawn) => {
-                self.process_evpn_withdraw_chunk(peer, evpn_withdrawn);
-            }
-            PendingRouteChunk::EvpnAnnounced(evpn_announced) => {
-                self.process_evpn_announce_chunk(peer, evpn_announced);
-            }
-        }
-
-        if pending.has_more() {
-            self.pending_route_batches.push_front(pending);
-        } else {
-            // Batch fully drained — distribute the changes accumulated
-            // across all its chunks in one coalesced outbound pass.
-            self.flush_pending_distribute();
-        }
-
-        true
-    }
-
-    /// Distribute the best-path changes accumulated across the chunks of a
-    /// route batch in a single pass, then clear the accumulator. Called when
-    /// a batch drains (see [`Self::process_next_route_chunk`]); a no-op when
-    /// nothing accumulated. Deferring distribution this way coalesces a
-    /// multi-chunk initial-load flood into one outbound batch per peer
-    /// instead of one per 1024-route chunk, while `recompute_best` still runs
-    /// per chunk so Loc-RIB and route events stay live mid-batch.
-    fn flush_pending_distribute(&mut self) {
-        if self.pending_distribute_changed.is_empty() && self.pending_distribute_affected.is_empty()
-        {
-            return;
-        }
-        let changed = std::mem::take(&mut self.pending_distribute_changed);
-        let affected = std::mem::take(&mut self.pending_distribute_affected);
-        self.distribute_changes(&changed, &affected);
-    }
-
-    fn process_withdraw_chunk(&mut self, peer: IpAddr, withdrawn: Vec<(Prefix, u32)>) {
-        let active_refresh = self
-            .refresh_in_progress
-            .get(&peer)
-            .cloned()
-            .unwrap_or_default();
-        let mut affected = HashSet::new();
-        let mut removed_stale_counts: HashMap<(Afi, Safi), usize> = HashMap::new();
-
-        let (rib_len, flowspec_len) = {
-            let rib = self
-                .ribs
-                .get_mut(&peer)
-                .expect("peer rib must exist before chunk processing");
-
-            for (prefix, path_id) in withdrawn {
-                if rib.withdraw(&prefix, path_id) {
-                    debug!(%peer, %prefix, path_id, "withdrawn");
-                    affected.insert(prefix);
-                }
-                let family = prefix_family(&prefix);
-                if active_refresh.contains(&family)
-                    && let Some(stale) = self.refresh_stale_routes.get_mut(&peer)
-                    && stale.remove(&(prefix, path_id))
-                {
-                    *removed_stale_counts.entry(family).or_default() += 1;
-                }
-            }
-            debug!(%peer, routes = rib.len(), "rib updated");
-            (rib.len(), rib.flowspec_len())
-        };
-
-        let peer_label = peer.to_string();
-        self.metrics
-            .set_rib_prefixes(&peer_label, "all", gauge_val(rib_len));
-        self.metrics
-            .set_rib_prefixes(&peer_label, "flowspec", gauge_val(flowspec_len));
-        for ((afi, safi), count) in removed_stale_counts {
-            self.decrement_refresh_stale_count(peer, afi, safi, count);
-        }
-        self.update_peer_refresh_metrics(peer);
-        if !affected.is_empty() {
-            // Recompute Loc-RIB now — best-path, route events, and
-            // partial-progress queries stay live mid-batch — but accumulate
-            // the distribution and flush it once when the batch drains, so a
-            // multi-chunk flood coalesces into one outbound pass per peer.
-            let changed = self.recompute_best(&affected);
-            self.pending_distribute_changed.extend(changed);
-            self.pending_distribute_affected.extend(affected);
-            if let Some(rib) = self.ribs.get_mut(&peer) {
-                rib.gc_intern_table();
-            }
-        }
-    }
-
-    pub(super) fn process_announce_chunk(
-        &mut self,
-        peer: IpAddr,
-        announced: Vec<crate::route::Route>,
-    ) {
-        let active_refresh = self
-            .refresh_in_progress
-            .get(&peer)
-            .cloned()
-            .unwrap_or_default();
-        let vrp_table: Option<Arc<VrpTable>> = self.vrp_table.as_ref().map(Arc::clone);
-        let mut affected = HashSet::new();
-        let mut removed_stale_counts: HashMap<(Afi, Safi), usize> = HashMap::new();
-        let mut any_replaced = false;
-
-        let (rib_len, flowspec_len) = {
-            let rib = self
-                .ribs
-                .get_mut(&peer)
-                .expect("peer rib must exist before chunk processing");
-
-            for mut route in announced {
-                if let Some(ref table) = vrp_table {
-                    route.validation_state = validate_route_rpki(&route, table);
-                }
-                if let Some(ref table) = self.aspa_table {
-                    route.aspa_state = validate_route_aspa(&route, table);
-                }
-                debug!(%peer, prefix = %route.prefix, "announced");
-                affected.insert(route.prefix);
-                let prefix = route.prefix;
-                let path_id = route.path_id;
-                any_replaced |= rib.insert(route);
-                let family = prefix_family(&prefix);
-                if active_refresh.contains(&family)
-                    && let Some(stale) = self.refresh_stale_routes.get_mut(&peer)
-                    && stale.remove(&(prefix, path_id))
-                {
-                    *removed_stale_counts.entry(family).or_default() += 1;
-                }
-            }
-            debug!(%peer, routes = rib.len(), "rib updated");
-            (rib.len(), rib.flowspec_len())
-        };
-
-        let peer_label = peer.to_string();
-        self.metrics
-            .set_rib_prefixes(&peer_label, "all", gauge_val(rib_len));
-        self.metrics
-            .set_rib_prefixes(&peer_label, "flowspec", gauge_val(flowspec_len));
-        for ((afi, safi), count) in removed_stale_counts {
-            self.decrement_refresh_stale_count(peer, afi, safi, count);
-        }
-        self.update_peer_refresh_metrics(peer);
-        if !affected.is_empty() {
-            // Recompute Loc-RIB now — best-path, route events, and
-            // partial-progress queries stay live mid-batch — but accumulate
-            // the distribution and flush it once when the batch drains, so a
-            // multi-chunk flood coalesces into one outbound pass per peer.
-            let changed = self.recompute_best(&affected);
-            self.pending_distribute_changed.extend(changed);
-            self.pending_distribute_affected.extend(affected);
-            if any_replaced && let Some(rib) = self.ribs.get_mut(&peer) {
-                rib.gc_intern_table();
-            }
-        }
-    }
-
-    fn process_flowspec_withdraw_chunk(
-        &mut self,
-        peer: IpAddr,
-        flowspec_withdrawn: Vec<FlowSpecRule>,
-    ) {
-        let active_refresh = self
-            .refresh_in_progress
-            .get(&peer)
-            .cloned()
-            .unwrap_or_default();
-        let mut fs_affected: HashSet<FlowSpecRule> = HashSet::new();
-        let mut removed_stale_counts: HashMap<(Afi, Safi), usize> = HashMap::new();
-
-        let (rib_len, flowspec_len) = {
-            let rib = self
-                .ribs
-                .get_mut(&peer)
-                .expect("peer rib must exist before chunk processing");
-
-            for rule in flowspec_withdrawn {
-                if rib.withdraw_flowspec(&rule, 0) {
-                    debug!(%peer, rule = %rule, "flowspec withdrawn");
-                    fs_affected.insert(rule.clone());
-                }
-                if active_refresh.iter().any(|(afi, safi)| {
-                    *safi == Safi::FlowSpec && matches!(afi, Afi::Ipv4 | Afi::Ipv6)
-                }) && let Some(stale) = self.refresh_stale_flowspec.get_mut(&peer)
-                {
-                    stale.retain(|(stale_afi, stale_rule, _)| {
-                        let keep = stale_rule != &rule;
-                        if !keep {
-                            *removed_stale_counts
-                                .entry((*stale_afi, Safi::FlowSpec))
-                                .or_default() += 1;
-                        }
-                        keep
-                    });
-                }
-            }
-
-            debug!(%peer, routes = rib.len(), "rib updated");
-            (rib.len(), rib.flowspec_len())
-        };
-
-        let peer_label = peer.to_string();
-        self.metrics
-            .set_rib_prefixes(&peer_label, "all", gauge_val(rib_len));
-        self.metrics
-            .set_rib_prefixes(&peer_label, "flowspec", gauge_val(flowspec_len));
-        for ((afi, safi), count) in removed_stale_counts {
-            self.decrement_refresh_stale_count(peer, afi, safi, count);
-        }
-        self.update_peer_refresh_metrics(peer);
-        if !fs_affected.is_empty() {
-            self.recompute_and_distribute_flowspec(&fs_affected);
-        }
-    }
-
-    pub(super) fn process_evpn_withdraw_chunk(
-        &mut self,
-        peer: IpAddr,
-        evpn_withdrawn: Vec<rustbgpd_wire::EvpnRouteKey>,
-    ) {
-        let active_refresh = self
-            .refresh_in_progress
-            .get(&peer)
-            .cloned()
-            .unwrap_or_default();
-        let evpn_refresh_active = active_refresh.contains(&(Afi::L2Vpn, Safi::Evpn));
-        let mut affected: HashSet<rustbgpd_wire::EvpnRouteKey> = HashSet::new();
-        let mut removed_stale_count = 0usize;
-        let evpn_len = {
-            let rib = self
-                .ribs
-                .get_mut(&peer)
-                .expect("peer rib must exist before chunk processing");
-            for key in evpn_withdrawn {
-                if rib.withdraw_evpn(&key) {
-                    debug!(%peer, route_type = key.route_type(), "evpn withdrawn");
-                    affected.insert(key);
-                    if evpn_refresh_active
-                        && let Some(stale) = self.refresh_stale_evpn.get_mut(&peer)
-                        && stale.remove(&key)
-                    {
-                        removed_stale_count += 1;
-                    }
-                }
-            }
-            rib.evpn_len()
-        };
-        self.metrics
-            .set_rib_prefixes(&peer.to_string(), "evpn", gauge_val(evpn_len));
-        self.decrement_refresh_stale_count(peer, Afi::L2Vpn, Safi::Evpn, removed_stale_count);
-        self.update_peer_refresh_metrics(peer);
-        if !affected.is_empty() {
-            self.recompute_and_distribute_evpn(&affected);
-            if let Some(rib) = self.ribs.get_mut(&peer) {
-                rib.gc_intern_table();
-            }
-        }
-    }
-
-    pub(super) fn process_evpn_announce_chunk(
-        &mut self,
-        peer: IpAddr,
-        evpn_announced: Vec<crate::route::EvpnRibRoute>,
-    ) {
-        let active_refresh = self
-            .refresh_in_progress
-            .get(&peer)
-            .cloned()
-            .unwrap_or_default();
-        let evpn_refresh_active = active_refresh.contains(&(Afi::L2Vpn, Safi::Evpn));
-        let mut affected: HashSet<rustbgpd_wire::EvpnRouteKey> = HashSet::new();
-        let mut removed_stale_count = 0usize;
-        let mut any_replaced = false;
-        let evpn_len = {
-            let rib = self
-                .ribs
-                .get_mut(&peer)
-                .expect("peer rib must exist before chunk processing");
-            for route in evpn_announced {
-                debug!(%peer, route_type = route.route_type(), "evpn announced");
-                let key = route.key();
-                affected.insert(key);
-                any_replaced |= rib.insert_evpn(route);
-                // Enhanced Route Refresh: re-advertised key removes from
-                // the stale set so EoRR's sweep doesn't withdraw it.
-                if evpn_refresh_active
-                    && let Some(stale) = self.refresh_stale_evpn.get_mut(&peer)
-                    && stale.remove(&key)
-                {
-                    removed_stale_count += 1;
-                }
-            }
-            rib.evpn_len()
-        };
-        self.metrics
-            .set_rib_prefixes(&peer.to_string(), "evpn", gauge_val(evpn_len));
-        self.decrement_refresh_stale_count(peer, Afi::L2Vpn, Safi::Evpn, removed_stale_count);
-        self.update_peer_refresh_metrics(peer);
-        if !affected.is_empty() {
-            self.recompute_and_distribute_evpn(&affected);
-            if any_replaced && let Some(rib) = self.ribs.get_mut(&peer) {
-                rib.gc_intern_table();
-            }
-        }
-    }
-
-    fn process_flowspec_announce_chunk(
-        &mut self,
-        peer: IpAddr,
-        flowspec_announced: Vec<crate::route::FlowSpecRoute>,
-    ) {
-        let active_refresh = self
-            .refresh_in_progress
-            .get(&peer)
-            .cloned()
-            .unwrap_or_default();
-        let mut fs_affected: HashSet<FlowSpecRule> = HashSet::new();
-        let mut removed_stale_counts: HashMap<(Afi, Safi), usize> = HashMap::new();
-
-        let (rib_len, flowspec_len) = {
-            let rib = self
-                .ribs
-                .get_mut(&peer)
-                .expect("peer rib must exist before chunk processing");
-
-            for route in flowspec_announced {
-                debug!(%peer, rule = %route.rule, "flowspec announced");
-                let stale_key = (route.afi, route.rule.clone(), route.path_id);
-                fs_affected.insert(route.rule.clone());
-                rib.insert_flowspec(route);
-                if active_refresh.contains(&(stale_key.0, Safi::FlowSpec))
-                    && let Some(stale) = self.refresh_stale_flowspec.get_mut(&peer)
-                    && stale.remove(&stale_key)
-                {
-                    *removed_stale_counts
-                        .entry((stale_key.0, Safi::FlowSpec))
-                        .or_default() += 1;
-                }
-            }
-
-            debug!(%peer, routes = rib.len(), "rib updated");
-            (rib.len(), rib.flowspec_len())
-        };
-
-        let peer_label = peer.to_string();
-        self.metrics
-            .set_rib_prefixes(&peer_label, "all", gauge_val(rib_len));
-        self.metrics
-            .set_rib_prefixes(&peer_label, "flowspec", gauge_val(flowspec_len));
-        for ((afi, safi), count) in removed_stale_counts {
-            self.decrement_refresh_stale_count(peer, afi, safi, count);
-        }
-        self.update_peer_refresh_metrics(peer);
-        if !fs_affected.is_empty() {
-            self.recompute_and_distribute_flowspec(&fs_affected);
-        }
-    }
-
-    /// Recompute Loc-RIB best path for a set of affected prefixes.
-    /// Returns the set of prefixes that actually changed.
-    /// Also emits route events to the broadcast channel.
-    pub(super) fn recompute_best(&mut self, affected: &HashSet<Prefix>) -> HashSet<Prefix> {
-        let mut changed = HashSet::new();
-        for prefix in affected {
-            let previous_best = self.loc_rib.get(prefix).map(|r| (r.peer, r.path_id));
-            let candidates = self.ribs.values().flat_map(|rib| rib.iter_prefix(prefix));
-            let did_change = self.loc_rib.recompute(*prefix, candidates);
-            if did_change {
-                changed.insert(*prefix);
-                let current_best = self.loc_rib.get(prefix);
-                match (previous_best, current_best) {
-                    (None, Some(best)) => {
-                        debug!(%prefix, peer = %best.peer, "best path added");
-                        self.publish_route_event(RouteEvent {
-                            event_id: 0,
-                            event_type: RouteEventType::Added,
-                            prefix: *prefix,
-                            peer: Some(best.peer),
-                            previous_peer: None,
-                            target_peer: None,
-                            timestamp: crate::event::unix_timestamp_now(),
-                            path_id: best.path_id,
-                            reason: String::new(),
-                        });
-                    }
-                    (Some((old_peer, old_path_id)), None) => {
-                        debug!(%prefix, "best path removed");
-                        self.publish_route_event(RouteEvent {
-                            event_id: 0,
-                            event_type: RouteEventType::Withdrawn,
-                            prefix: *prefix,
-                            peer: None,
-                            previous_peer: Some(old_peer),
-                            target_peer: None,
-                            timestamp: crate::event::unix_timestamp_now(),
-                            path_id: old_path_id,
-                            reason: String::new(),
-                        });
-                    }
-                    (Some((old_peer, _old_path_id)), Some(best)) => {
-                        debug!(%prefix, peer = %best.peer, "best path changed");
-                        self.publish_route_event(RouteEvent {
-                            event_id: 0,
-                            event_type: RouteEventType::BestChanged,
-                            prefix: *prefix,
-                            peer: Some(best.peer),
-                            previous_peer: Some(old_peer),
-                            target_peer: None,
-                            timestamp: crate::event::unix_timestamp_now(),
-                            path_id: best.path_id,
-                            reason: String::new(),
-                        });
-                    }
-                    (None, None) => {}
-                }
-            }
-        }
-        self.metrics
-            .set_loc_rib_prefixes("all", gauge_val(self.loc_rib.len()));
-        changed
-    }
-
-    pub(super) fn handle_inject_route(
-        &mut self,
-        route: crate::route::Route,
-        reply: tokio::sync::oneshot::Sender<Result<(), RibCommandError>>,
-    ) {
-        let prefix = route.prefix;
-        let rib = self
-            .ribs
-            .entry(LOCAL_PEER)
-            .or_insert_with(|| AdjRibIn::new(LOCAL_PEER));
-        let replaced = rib.insert(route);
-        debug!(%prefix, "injected local route");
-        self.metrics
-            .set_rib_prefixes(&LOCAL_PEER.to_string(), "all", gauge_val(rib.len()));
-
-        let mut affected = HashSet::new();
-        affected.insert(prefix);
-        let changed = self.recompute_best(&affected);
-        self.distribute_changes(&changed, &affected);
-        if replaced && let Some(rib) = self.ribs.get_mut(&LOCAL_PEER) {
-            rib.gc_intern_table();
-        }
-
-        let _ = reply.send(Ok(()));
-    }
-
-    pub(super) fn handle_withdraw_injected(
-        &mut self,
-        prefix: Prefix,
-        path_id: u32,
-        reply: tokio::sync::oneshot::Sender<Result<(), RibCommandError>>,
-    ) {
-        let rib = self
-            .ribs
-            .entry(LOCAL_PEER)
-            .or_insert_with(|| AdjRibIn::new(LOCAL_PEER));
-        if rib.withdraw(&prefix, path_id) {
-            debug!(%prefix, "withdrawn injected route");
-            self.metrics
-                .set_rib_prefixes(&LOCAL_PEER.to_string(), "all", gauge_val(rib.len()));
-            let mut affected = HashSet::new();
-            affected.insert(prefix);
-            let changed = self.recompute_best(&affected);
-            self.distribute_changes(&changed, &affected);
-            if let Some(rib) = self.ribs.get_mut(&LOCAL_PEER) {
-                rib.gc_intern_table();
-            }
-            let _ = reply.send(Ok(()));
-        } else {
-            let _ = reply.send(Err(RibCommandError::not_found(format!(
-                "prefix {prefix} not found"
-            ))));
-        }
-    }
-
-    pub(super) fn handle_inject_flowspec(
-        &mut self,
-        route: crate::route::FlowSpecRoute,
-        reply: tokio::sync::oneshot::Sender<Result<(), RibCommandError>>,
-    ) {
-        let rule = route.rule.clone();
-        let rib = self
-            .ribs
-            .entry(LOCAL_PEER)
-            .or_insert_with(|| AdjRibIn::new(LOCAL_PEER));
-        rib.insert_flowspec(route);
-        debug!(rule = %rule, "injected local FlowSpec route");
-        let mut fs_affected = HashSet::new();
-        fs_affected.insert(rule);
-        self.recompute_and_distribute_flowspec(&fs_affected);
-        let _ = reply.send(Ok(()));
-    }
-
-    pub(super) fn handle_withdraw_flowspec(
-        &mut self,
-        rule: FlowSpecRule,
-        reply: tokio::sync::oneshot::Sender<Result<(), RibCommandError>>,
-    ) {
-        let rib = self
-            .ribs
-            .entry(LOCAL_PEER)
-            .or_insert_with(|| AdjRibIn::new(LOCAL_PEER));
-        if rib.withdraw_flowspec(&rule, 0) {
-            debug!(rule = %rule, "withdrawn injected FlowSpec route");
-            let mut fs_affected = HashSet::new();
-            fs_affected.insert(rule);
-            self.recompute_and_distribute_flowspec(&fs_affected);
-            let _ = reply.send(Ok(()));
-        } else {
-            let _ = reply.send(Err(RibCommandError::not_found(format!(
-                "FlowSpec rule {rule} not found"
-            ))));
-        }
-    }
-
-    pub(super) fn handle_inject_evpn(
-        &mut self,
-        route: crate::route::EvpnRibRoute,
-        reply: tokio::sync::oneshot::Sender<Result<(), RibCommandError>>,
-    ) {
-        let key = route.key();
-        let rib = self
-            .ribs
-            .entry(LOCAL_PEER)
-            .or_insert_with(|| AdjRibIn::new(LOCAL_PEER));
-        let replaced = rib.insert_evpn(route);
-        debug!(?key, "injected local EVPN route");
-        let mut evpn_affected = HashSet::new();
-        evpn_affected.insert(key);
-        self.recompute_and_distribute_evpn(&evpn_affected);
-        if replaced && let Some(rib) = self.ribs.get_mut(&LOCAL_PEER) {
-            rib.gc_intern_table();
-        }
-        let _ = reply.send(Ok(()));
-    }
-
-    pub(super) fn handle_withdraw_evpn(
-        &mut self,
-        key: rustbgpd_wire::EvpnRouteKey,
-        reply: tokio::sync::oneshot::Sender<Result<(), RibCommandError>>,
-    ) {
-        let rib = self
-            .ribs
-            .entry(LOCAL_PEER)
-            .or_insert_with(|| AdjRibIn::new(LOCAL_PEER));
-        if rib.withdraw_evpn(&key) {
-            debug!(?key, "withdrawn injected EVPN route");
-            let mut evpn_affected = HashSet::new();
-            evpn_affected.insert(key);
-            self.recompute_and_distribute_evpn(&evpn_affected);
-            if let Some(rib) = self.ribs.get_mut(&LOCAL_PEER) {
-                rib.gc_intern_table();
-            }
-            let _ = reply.send(Ok(()));
-        } else {
-            let _ = reply.send(Err(RibCommandError::not_found(format!(
-                "EVPN route key {key:?} not found"
-            ))));
-        }
-    }
-
-    /// Multi-path distribution for a single prefix to a single peer.
-    ///
-    /// Collects all candidates from all Adj-RIB-In entries, filters by
-    /// split-horizon/iBGP/family/policy, sorts by best-path, takes top N,
-    /// and diffs against `AdjRibOut` to produce announces and withdrawals.
-    #[expect(
-        clippy::too_many_arguments,
-        clippy::too_many_lines,
-        reason = "multipath export keeps peer, policy, and Adj-RIB-Out diff state together"
-    )]
-    pub(super) fn distribute_multipath_prefix(
-        ribs: &HashMap<IpAddr, AdjRibIn>,
-        rib_out: &AdjRibOut,
-        peer_is_rr_client: &HashMap<IpAddr, bool>,
-        prefix: &Prefix,
-        target_peer: IpAddr,
-        target_peer_asn: Option<u32>,
-        target_peer_group: Option<&str>,
-        send_max: u32,
-        target_is_ebgp: bool,
-        target_is_rr_client: bool,
-        cluster_id: Option<Ipv4Addr>,
-        sendable: Option<&Vec<(Afi, Safi)>>,
-        export_pol: Option<&PolicyChain>,
-        orf_filter: Option<&crate::orf::OrfFilterSet>,
-        orr: Option<(&crate::orr::OrrTopology, &crate::orr::SpfResult)>,
-        metrics: &BgpMetrics,
-        policy_stats: &mut NeighborPolicyStats,
-        target_peer_label: &str,
-        announce: &mut Vec<crate::route::Route>,
-        withdraw: &mut Vec<(Prefix, u32)>,
-        nh_override_flags: &mut Vec<Option<rustbgpd_policy::NextHopAction>>,
-        policy_filtered: &mut Vec<PolicyFilteredRouteKey>,
-        force: bool,
-    ) {
-        use crate::best_path::{best_path_cmp, best_path_cmp_orr};
-
-        // Sendable family check
-        let family = match prefix {
-            Prefix::V4(_) => (Afi::Ipv4, Safi::Unicast),
-            Prefix::V6(_) => (Afi::Ipv6, Safi::Unicast),
-        };
-        if !sendable.is_some_and(|f| f.contains(&family)) {
-            // Withdraw all previously advertised paths for this prefix
-            for &path_id in rib_out.path_ids_for_prefix(prefix) {
-                withdraw.push((*prefix, path_id));
-            }
-            return;
-        }
-
-        // ORF check (RFC 5291): filters the prefix, not individual Add-Path
-        // path-ids — gate the whole prefix once, before collecting candidates.
-        if orf_filter.is_some_and(|f| !f.permits(prefix)) {
-            for &path_id in rib_out.path_ids_for_prefix(prefix) {
-                withdraw.push((*prefix, path_id));
-            }
-            return;
-        }
-
-        // Collect all candidates across all Adj-RIB-In entries
-        let mut candidates: Vec<&crate::route::Route> = ribs
-            .values()
-            .flat_map(|rib| rib.iter_prefix(prefix))
-            .filter(|route| {
-                // Split horizon: exclude routes from the target peer
-                if route.peer == target_peer {
-                    return false;
-                }
-                // iBGP split-horizon / RFC 4456 reflection
-                if should_suppress_ibgp_inner(
-                    route,
-                    target_is_ebgp,
-                    target_is_rr_client,
-                    cluster_id,
-                    peer_is_rr_client,
-                ) {
-                    return false;
-                }
-                true
-            })
-            .collect();
-
-        // Sort by best-path preference (best first). A target bound to a
-        // resolved ORR vantage ranks by the vantage's interior cost to
-        // each NEXT_HOP first (RFC 9107 §3.1) — comparator swap only.
-        match orr {
-            Some((topology, spf)) => candidates.sort_by(|a, b| {
-                best_path_cmp_orr(
-                    a,
-                    b,
-                    spf.cost_to(topology, a.next_hop),
-                    spf.cost_to(topology, b.next_hop),
-                )
-            }),
-            None => candidates.sort_by(|a, b| best_path_cmp(a, b)),
-        }
-
-        // Walk candidates, evaluate export policy, assign path_ids 1..N
-        let needs_as_path_string = export_pol.is_some_and(PolicyChain::requires_as_path_string);
-        let mut next_rank: u32 = 1;
-        let limit = if send_max == u32::MAX {
-            usize::MAX
-        } else {
-            send_max as usize
-        };
-        for candidate in &candidates {
-            if (next_rank as usize) > limit {
-                break;
-            }
-
-            // Export policy check per-candidate
-            let aspath_str = if needs_as_path_string {
-                candidate
-                    .as_path()
-                    .map_or_else(String::new, rustbgpd_wire::AsPath::to_aspath_string)
-            } else {
-                String::new()
-            };
-            let aspath_len = candidate.as_path().map_or(0, rustbgpd_wire::AsPath::len);
-            let ctx = RouteContext {
-                prefix: Some(*prefix),
-                next_hop: Some(candidate.next_hop),
-                extended_communities: candidate.extended_communities(),
-                communities: candidate.communities(),
-                large_communities: candidate.large_communities(),
-                as_path_str: &aspath_str,
-                as_path_len: aspath_len,
-                validation_state: candidate.validation_state,
-                aspa_state: candidate.aspa_state,
-                peer_address: Some(target_peer),
-                peer_asn: target_peer_asn,
-                peer_group: target_peer_group,
-                route_type: Some(route_type(candidate.origin_type)),
-                evpn_route_type: None,
-                local_pref: candidate.local_pref_attr(),
-                med: candidate.med_attr(),
-            };
-            let (result, evaluation) = evaluate_chain_with_attribution(export_pol, &ctx);
-            record_export_policy_eval(metrics, policy_stats, target_peer_label, &evaluation);
-            if result.action != PolicyAction::Permit {
-                policy_filtered.push(PolicyFilteredRouteKey {
-                    target_peer,
-                    source_peer: candidate.peer,
-                    prefix: *prefix,
-                    path_id: candidate.path_id,
-                });
-                continue;
-            }
-
-            // Apply export modifications — skip deep clone when no mods needed
-            let mut modified = (*candidate).clone();
-            let nh_action = if result.modifications.is_empty() {
-                None
-            } else {
-                let nh = rustbgpd_policy::apply_modifications(
-                    std::sync::Arc::make_mut(&mut modified.attributes),
-                    &result.modifications,
-                );
-                if let Some(rustbgpd_policy::NextHopAction::Specific(addr)) = &nh {
-                    modified.next_hop = *addr;
-                }
-                nh
-            };
-            modified.path_id = next_rank;
-
-            // Only announce if different from what's already in AdjRibOut.
-            // `force` mode bypasses the equality check — see
-            // `distribute_single_best_prefix` for the GShut rationale.
-            let changed = force
-                || rib_out
-                    .get(prefix, next_rank)
-                    .is_none_or(|existing| !routes_equal(existing, &modified));
-            if changed {
-                nh_override_flags.push(nh_action);
-                announce.push(modified);
-            }
-
-            next_rank += 1;
-        }
-
-        // Withdraw any previously advertised path_ids beyond the new set
-        for &path_id in rib_out.path_ids_for_prefix(prefix) {
-            if path_id >= next_rank {
-                withdraw.push((*prefix, path_id));
-            }
-        }
-    }
-
-    #[expect(
-        clippy::too_many_arguments,
-        clippy::too_many_lines,
-        reason = "single-best export keeps peer, policy, and Adj-RIB-Out diff state together"
-    )]
-    pub(super) fn distribute_single_best_prefix(
-        loc_rib: &LocRib,
-        rib_out: &AdjRibOut,
-        peer_is_rr_client: &HashMap<IpAddr, bool>,
-        prefix: &Prefix,
-        target_peer: IpAddr,
-        target_peer_asn: Option<u32>,
-        target_peer_group: Option<&str>,
-        target_is_ebgp: bool,
-        target_is_rr_client: bool,
-        cluster_id: Option<Ipv4Addr>,
-        sendable: Option<&Vec<(Afi, Safi)>>,
-        export_pol: Option<&PolicyChain>,
-        orf_filter: Option<&crate::orf::OrfFilterSet>,
-        metrics: &BgpMetrics,
-        policy_stats: &mut NeighborPolicyStats,
-        target_peer_label: &str,
-        announce: &mut Vec<crate::route::Route>,
-        withdraw: &mut Vec<(Prefix, u32)>,
-        nh_override_flags: &mut Vec<Option<rustbgpd_policy::NextHopAction>>,
-        policy_filtered: &mut Vec<PolicyFilteredRouteKey>,
-        force: bool,
-    ) {
-        let existing_path_ids = rib_out.path_ids_for_prefix(prefix);
-
-        let Some(best) = loc_rib.get(prefix) else {
-            for &path_id in existing_path_ids {
-                withdraw.push((*prefix, path_id));
-            }
-            return;
-        };
-
-        // Split horizon: don't send route back to its source
-        if best.peer == target_peer {
-            for &path_id in existing_path_ids {
-                withdraw.push((*prefix, path_id));
-            }
-            return;
-        }
-
-        // iBGP split-horizon / RFC 4456 reflection rules
-        if should_suppress_ibgp_inner(
-            best,
-            target_is_ebgp,
-            target_is_rr_client,
-            cluster_id,
-            peer_is_rr_client,
-        ) {
-            for &path_id in existing_path_ids {
-                withdraw.push((*prefix, path_id));
-            }
-            return;
-        }
-
-        // Sendable family check
-        let family = prefix_family(prefix);
-        if !sendable.is_some_and(|f| f.contains(&family)) {
-            for &path_id in existing_path_ids {
-                withdraw.push((*prefix, path_id));
-            }
-            return;
-        }
-
-        // Outbound Route Filter check (RFC 5291): a peer-pushed prefix filter
-        // applied *before* export policy. ORF denial is a silent withdraw —
-        // not a policy denial — so it is deliberately not recorded in the
-        // `policy_filtered` set or the export-policy counters.
-        if orf_filter.is_some_and(|f| !f.permits(prefix)) {
-            for &path_id in existing_path_ids {
-                withdraw.push((*prefix, path_id));
-            }
-            return;
-        }
-
-        // Export policy check
-        let aspath_str = if export_pol.is_some_and(PolicyChain::requires_as_path_string) {
-            best.as_path()
-                .map_or_else(String::new, rustbgpd_wire::AsPath::to_aspath_string)
-        } else {
-            String::new()
-        };
-        let aspath_len = best.as_path().map_or(0, rustbgpd_wire::AsPath::len);
-        let ctx = RouteContext {
-            prefix: Some(*prefix),
-            next_hop: Some(best.next_hop),
-            extended_communities: best.extended_communities(),
-            communities: best.communities(),
-            large_communities: best.large_communities(),
-            as_path_str: &aspath_str,
-            as_path_len: aspath_len,
-            validation_state: best.validation_state,
-            aspa_state: best.aspa_state,
-            peer_address: Some(target_peer),
-            peer_asn: target_peer_asn,
-            peer_group: target_peer_group,
-            route_type: Some(route_type(best.origin_type)),
-            evpn_route_type: None,
-            local_pref: best.local_pref_attr(),
-            med: best.med_attr(),
-        };
-        let (result, evaluation) = evaluate_chain_with_attribution(export_pol, &ctx);
-        record_export_policy_eval(metrics, policy_stats, target_peer_label, &evaluation);
-        if result.action != PolicyAction::Permit {
-            policy_filtered.push(PolicyFilteredRouteKey {
-                target_peer,
-                source_peer: best.peer,
-                prefix: *prefix,
-                path_id: best.path_id,
-            });
-            for &path_id in existing_path_ids {
-                withdraw.push((*prefix, path_id));
-            }
-            return;
-        }
-
-        // Apply export modifications to a clone — skip the deep clone of
-        // the Arc<Vec<PathAttribute>> when no modifications are needed.
-        let mut modified = best.clone();
-        let nh_action = if result.modifications.is_empty() {
-            None
-        } else {
-            let nh = rustbgpd_policy::apply_modifications(
-                std::sync::Arc::make_mut(&mut modified.attributes),
-                &result.modifications,
-            );
-            if let Some(rustbgpd_policy::NextHopAction::Specific(addr)) = &nh {
-                modified.next_hop = *addr;
-            }
-            nh
-        };
-        modified.path_id = 0;
-
-        // `force` mode: bypass the AdjRibOut equality suppression so
-        // currently-advertised routes re-emit even when the RIB-level
-        // attribute set is unchanged. Used by RFC 8326 GShut toggle
-        // where the wire change is applied in transport AFTER this
-        // diff. Without `force` the dirty-peers resync would skip
-        // every route that already lives in AdjRibOut, defeating the
-        // whole point of `RefreshPeerOutbound`.
-        let changed = force
-            || rib_out
-                .get(prefix, 0)
-                .is_none_or(|existing| !routes_equal(existing, &modified));
-        if changed {
-            nh_override_flags.push(nh_action);
-            announce.push(modified);
-        }
-
-        // Clean up any stale multi-path entries if this prefix was previously
-        // advertised via Add-Path and is now single-best.
-        for &path_id in existing_path_ids {
-            if path_id != 0 {
-                withdraw.push((*prefix, path_id));
-            }
-        }
-    }
-
-    /// RFC 9107 ORR single-best distribution for one prefix to one peer
-    /// bound to a *resolved* vantage.
-    ///
-    /// The best is NOT the Loc-RIB best: the candidate set is collected
-    /// and filtered per target peer exactly like
-    /// [`Self::distribute_multipath_prefix`] (all Adj-RIB-Ins, split
-    /// horizon, iBGP/RR suppression), then the winner is picked with
-    /// [`crate::best_path::best_path_cmp_orr`] using the vantage's SPF
-    /// cost to each candidate's `NEXT_HOP`. The export tail (policy,
-    /// modifications, Adj-RIB-Out diff) mirrors
-    /// [`Self::distribute_single_best_prefix`].
-    ///
-    /// Deliberately NO per-(vantage, prefix) winner memo — it would be
-    /// unsound: the candidate set is per-target-peer (split horizon and
-    /// RR suppression drop different routes for different targets), so a
-    /// memoized winner could be a route the next target must never
-    /// receive. The `SpfResult`'s internal NH-cost LRU is the sound
-    /// cache.
-    #[expect(
-        clippy::too_many_arguments,
-        clippy::too_many_lines,
-        reason = "ORR export keeps peer, policy, and Adj-RIB-Out diff state together"
-    )]
-    pub(super) fn distribute_orr_best_prefix(
-        ribs: &HashMap<IpAddr, AdjRibIn>,
-        rib_out: &AdjRibOut,
-        peer_is_rr_client: &HashMap<IpAddr, bool>,
-        orr_topology: &crate::orr::OrrTopology,
-        orr_spf: &crate::orr::SpfResult,
-        prefix: &Prefix,
-        target_peer: IpAddr,
-        target_peer_asn: Option<u32>,
-        target_peer_group: Option<&str>,
-        target_is_ebgp: bool,
-        target_is_rr_client: bool,
-        cluster_id: Option<Ipv4Addr>,
-        sendable: Option<&Vec<(Afi, Safi)>>,
-        export_pol: Option<&PolicyChain>,
-        orf_filter: Option<&crate::orf::OrfFilterSet>,
-        metrics: &BgpMetrics,
-        policy_stats: &mut NeighborPolicyStats,
-        target_peer_label: &str,
-        announce: &mut Vec<crate::route::Route>,
-        withdraw: &mut Vec<(Prefix, u32)>,
-        nh_override_flags: &mut Vec<Option<rustbgpd_policy::NextHopAction>>,
-        policy_filtered: &mut Vec<PolicyFilteredRouteKey>,
-        force: bool,
-    ) {
-        use crate::best_path::best_path_cmp_orr;
-
-        let existing_path_ids = rib_out.path_ids_for_prefix(prefix);
-
-        // Sendable family check
-        let family = prefix_family(prefix);
-        if !sendable.is_some_and(|f| f.contains(&family)) {
-            for &path_id in existing_path_ids {
-                withdraw.push((*prefix, path_id));
-            }
-            return;
-        }
-
-        // Outbound Route Filter check (RFC 5291): silent withdraw, not a
-        // policy denial — see `distribute_single_best_prefix`.
-        if orf_filter.is_some_and(|f| !f.permits(prefix)) {
-            for &path_id in existing_path_ids {
-                withdraw.push((*prefix, path_id));
-            }
-            return;
-        }
-
-        // Collect all candidates across all Adj-RIB-In entries —
-        // verbatim the multipath collector's per-target filter set.
-        let candidates = ribs
-            .values()
-            .flat_map(|rib| rib.iter_prefix(prefix))
-            .filter(|route| {
-                // Split horizon: exclude routes from the target peer
-                if route.peer == target_peer {
-                    return false;
-                }
-                // iBGP split-horizon / RFC 4456 reflection
-                if should_suppress_ibgp_inner(
-                    route,
-                    target_is_ebgp,
-                    target_is_rr_client,
-                    cluster_id,
-                    peer_is_rr_client,
-                ) {
-                    return false;
-                }
-                true
-            });
-
-        // Per-vantage best (RFC 9107): the vantage's interior cost to
-        // each NEXT_HOP breaks ties between step 5 and step 5.5.
-        let Some(best) = candidates.min_by(|a, b| {
-            best_path_cmp_orr(
-                a,
-                b,
-                orr_spf.cost_to(orr_topology, a.next_hop),
-                orr_spf.cost_to(orr_topology, b.next_hop),
-            )
-        }) else {
-            for &path_id in existing_path_ids {
-                withdraw.push((*prefix, path_id));
-            }
-            return;
-        };
-
-        // Export policy check — same tail as `distribute_single_best_prefix`.
-        let aspath_str = if export_pol.is_some_and(PolicyChain::requires_as_path_string) {
-            best.as_path()
-                .map_or_else(String::new, rustbgpd_wire::AsPath::to_aspath_string)
-        } else {
-            String::new()
-        };
-        let aspath_len = best.as_path().map_or(0, rustbgpd_wire::AsPath::len);
-        let ctx = RouteContext {
-            prefix: Some(*prefix),
-            next_hop: Some(best.next_hop),
-            extended_communities: best.extended_communities(),
-            communities: best.communities(),
-            large_communities: best.large_communities(),
-            as_path_str: &aspath_str,
-            as_path_len: aspath_len,
-            validation_state: best.validation_state,
-            aspa_state: best.aspa_state,
-            peer_address: Some(target_peer),
-            peer_asn: target_peer_asn,
-            peer_group: target_peer_group,
-            route_type: Some(route_type(best.origin_type)),
-            evpn_route_type: None,
-            local_pref: best.local_pref_attr(),
-            med: best.med_attr(),
-        };
-        let (result, evaluation) = evaluate_chain_with_attribution(export_pol, &ctx);
-        record_export_policy_eval(metrics, policy_stats, target_peer_label, &evaluation);
-        if result.action != PolicyAction::Permit {
-            policy_filtered.push(PolicyFilteredRouteKey {
-                target_peer,
-                source_peer: best.peer,
-                prefix: *prefix,
-                path_id: best.path_id,
-            });
-            for &path_id in existing_path_ids {
-                withdraw.push((*prefix, path_id));
-            }
-            return;
-        }
-
-        // Apply export modifications to a clone — skip the deep clone of
-        // the Arc<Vec<PathAttribute>> when no modifications are needed.
-        let mut modified = best.clone();
-        let nh_action = if result.modifications.is_empty() {
-            None
-        } else {
-            let nh = rustbgpd_policy::apply_modifications(
-                std::sync::Arc::make_mut(&mut modified.attributes),
-                &result.modifications,
-            );
-            if let Some(rustbgpd_policy::NextHopAction::Specific(addr)) = &nh {
-                modified.next_hop = *addr;
-            }
-            nh
-        };
-        modified.path_id = 0;
-
-        // `force` semantics as in `distribute_single_best_prefix`.
-        let changed = force
-            || rib_out
-                .get(prefix, 0)
-                .is_none_or(|existing| !routes_equal(existing, &modified));
-        if changed {
-            nh_override_flags.push(nh_action);
-            announce.push(modified);
-        }
-
-        // Clean up any stale multi-path entries if this prefix was
-        // previously advertised via Add-Path and is now single-best.
-        for &path_id in existing_path_ids {
-            if path_id != 0 {
-                withdraw.push((*prefix, path_id));
-            }
-        }
-    }
-
-    /// Stage `FlowSpec` announces and withdrawals for a set of rules.
-    ///
-    /// Uses `loc_rib` as the current best-route source and diffs against the
-    /// provided outbound view. Passing an empty `AdjRibOut` view causes a full
-    /// re-advertisement of the current `FlowSpec` export set, which is useful
-    /// for initial table dump and ROUTE-REFRESH responses.
-    #[expect(
-        clippy::too_many_arguments,
-        reason = "FlowSpec staging mirrors unicast distribution context"
-    )]
-    pub(super) fn stage_flowspec_rules(
-        loc_rib: &LocRib,
-        rib_out: &AdjRibOut,
-        peer_is_rr_client: &HashMap<IpAddr, bool>,
-        rules: &HashSet<FlowSpecRule>,
-        target_peer: IpAddr,
-        target_peer_asn: Option<u32>,
-        target_peer_group: Option<&str>,
-        target_is_ebgp: bool,
-        target_is_rr_client: bool,
-        cluster_id: Option<Ipv4Addr>,
-        sendable: Option<&Vec<(Afi, Safi)>>,
-        export_pol: Option<&PolicyChain>,
-        metrics: &BgpMetrics,
-        policy_stats: &mut NeighborPolicyStats,
-        target_peer_label: &str,
-        fs_announce: &mut Vec<crate::route::FlowSpecRoute>,
-        fs_withdraw: &mut Vec<FlowSpecRule>,
-    ) {
-        let needs_as_path_string = export_pol.is_some_and(PolicyChain::requires_as_path_string);
-        for rule in rules {
-            if let Some(best) = loc_rib.get_flowspec(rule) {
-                let fs_family = (best.afi, Safi::FlowSpec);
-                if !sendable.is_some_and(|f| f.contains(&fs_family)) {
-                    if rib_out.get_flowspec(rule).is_some() {
-                        fs_withdraw.push(rule.clone());
-                    }
-                    continue;
-                }
-
-                // Reuse the existing iBGP split-horizon / RR check. Only
-                // origin_type/peer/peer_router_id are relevant here.
-                if should_suppress_ibgp_inner(
-                    &crate::route::Route {
-                        prefix: Prefix::V4(rustbgpd_wire::Ipv4Prefix::new(
-                            Ipv4Addr::UNSPECIFIED,
-                            0,
-                        )),
-                        next_hop: best.peer,
-                        link_local_next_hop: None,
-                        next_hop_scope: None,
-                        peer: best.peer,
-                        attributes: std::sync::Arc::new(vec![]),
-                        received_at: best.received_at,
-                        origin_type: best.origin_type,
-                        peer_router_id: best.peer_router_id,
-                        is_stale: false,
-                        is_llgr_stale: false,
-                        path_id: 0,
-                        validation_state: rustbgpd_wire::RpkiValidation::NotFound,
-                        aspa_state: rustbgpd_wire::AspaValidation::Unknown,
-                        aspa_context: rustbgpd_wire::AspaValidationContext::default(),
-                    },
-                    target_is_ebgp,
-                    target_is_rr_client,
-                    cluster_id,
-                    peer_is_rr_client,
-                ) {
-                    if rib_out.get_flowspec(rule).is_some() {
-                        fs_withdraw.push(rule.clone());
-                    }
-                    continue;
-                }
-
-                let dest_prefix = best.rule.destination_prefix();
-                let prefix_for_policy = dest_prefix.unwrap_or(Prefix::V4(
-                    rustbgpd_wire::Ipv4Prefix::new(Ipv4Addr::UNSPECIFIED, 0),
-                ));
-                let aspath_str = if needs_as_path_string {
-                    best.as_path()
-                        .map_or_else(String::new, rustbgpd_wire::AsPath::to_aspath_string)
-                } else {
-                    String::new()
-                };
-                let aspath_len = best.as_path().map_or(0, rustbgpd_wire::AsPath::len);
-                let ctx = RouteContext {
-                    prefix: Some(prefix_for_policy),
-                    next_hop: None,
-                    extended_communities: best.extended_communities(),
-                    communities: best.communities(),
-                    large_communities: best.large_communities(),
-                    as_path_str: &aspath_str,
-                    as_path_len: aspath_len,
-                    validation_state: rustbgpd_wire::RpkiValidation::NotFound,
-                    aspa_state: rustbgpd_wire::AspaValidation::Unknown,
-                    peer_address: Some(target_peer),
-                    peer_asn: target_peer_asn,
-                    peer_group: target_peer_group,
-                    route_type: Some(route_type(best.origin_type)),
-                    evpn_route_type: None,
-                    local_pref: best.local_pref_attr(),
-                    med: best.med_attr(),
-                };
-                let (result, evaluation) =
-                    rustbgpd_policy::evaluate_chain_with_attribution(export_pol, &ctx);
-                record_export_policy_eval(metrics, policy_stats, target_peer_label, &evaluation);
-                if result.action == rustbgpd_policy::PolicyAction::Permit {
-                    fs_announce.push(best.clone());
-                } else if rib_out.get_flowspec(rule).is_some() {
-                    fs_withdraw.push(rule.clone());
-                }
-            } else if rib_out.get_flowspec(rule).is_some() {
-                fs_withdraw.push(rule.clone());
-            }
-        }
-    }
-
-    /// Stage EVPN announces and withdrawals for a set of affected keys.
-    ///
-    /// Selection semantics match `stage_flowspec_rules`: for each key, if the
-    /// `LocRib` has a best route, decide whether to advertise it to this peer
-    /// based on RFC 4456 iBGP split-horizon and export policy. Reflection
-    /// uses the same `should_suppress_ibgp_inner` helper with a synthetic
-    /// placeholder `Route` carrying the EVPN route's peer metadata.
-    ///
-    /// Phase 1: policy uses a placeholder `0.0.0.0/0` prefix; RT / community
-    /// matching works through the existing `RouteContext` fields.
-    #[expect(
-        clippy::too_many_arguments,
-        clippy::too_many_lines,
-        reason = "EVPN staging mirrors unicast distribution context for policy parity"
-    )]
-    pub(super) fn stage_evpn_routes(
-        loc_rib: &LocRib,
-        rib_out: &AdjRibOut,
-        peer_is_rr_client: &HashMap<IpAddr, bool>,
-        keys: &HashSet<rustbgpd_wire::EvpnRouteKey>,
-        target_peer: IpAddr,
-        target_peer_asn: Option<u32>,
-        target_peer_group: Option<&str>,
-        target_is_ebgp: bool,
-        target_is_rr_client: bool,
-        cluster_id: Option<Ipv4Addr>,
-        sendable: Option<&Vec<(Afi, Safi)>>,
-        export_pol: Option<&PolicyChain>,
-        metrics: &BgpMetrics,
-        policy_stats: &mut NeighborPolicyStats,
-        target_peer_label: &str,
-        evpn_announce: &mut Vec<crate::route::EvpnRibRoute>,
-        evpn_withdraw: &mut Vec<rustbgpd_wire::EvpnRouteKey>,
-        force: bool,
-    ) {
-        let needs_as_path_string = export_pol.is_some_and(PolicyChain::requires_as_path_string);
-        let evpn_family = (Afi::L2Vpn, Safi::Evpn);
-        let peer_supports_evpn = sendable.is_some_and(|f| f.contains(&evpn_family));
-
-        for key in keys {
-            if !peer_supports_evpn {
-                if rib_out.get_evpn(key).is_some() {
-                    evpn_withdraw.push(*key);
-                }
-                continue;
-            }
-
-            let Some(best) = loc_rib.get_evpn(key) else {
-                if rib_out.get_evpn(key).is_some() {
-                    evpn_withdraw.push(*key);
-                }
-                continue;
-            };
-
-            // Split horizon: don't send an EVPN route back to its source peer.
-            // Parallel to the unicast guard earlier in this module. Without
-            // this, an RR client could receive its own Type 2/3/4/etc. route
-            // back with ORIGINATOR_ID set to its own router-id — FRR and
-            // others drop such reflections, but RFC 4456 hygiene says we
-            // shouldn't emit them in the first place.
-            if best.peer == target_peer {
-                if rib_out.get_evpn(key).is_some() {
-                    evpn_withdraw.push(*key);
-                }
-                continue;
-            }
-
-            // Split-horizon / RR suppression check via a synthetic Route that
-            // carries only the fields should_suppress_ibgp_inner reads.
-            let probe = crate::route::Route {
-                prefix: Prefix::V4(rustbgpd_wire::Ipv4Prefix::new(Ipv4Addr::UNSPECIFIED, 0)),
-                next_hop: best.next_hop,
-                link_local_next_hop: None,
-                next_hop_scope: None,
-                peer: best.peer,
-                attributes: std::sync::Arc::new(vec![]),
-                received_at: best.received_at,
-                origin_type: best.origin_type,
-                peer_router_id: best.peer_router_id,
-                is_stale: false,
-                is_llgr_stale: false,
-                path_id: 0,
-                validation_state: rustbgpd_wire::RpkiValidation::NotFound,
-                aspa_state: rustbgpd_wire::AspaValidation::Unknown,
-                aspa_context: rustbgpd_wire::AspaValidationContext::default(),
-            };
-            if should_suppress_ibgp_inner(
-                &probe,
-                target_is_ebgp,
-                target_is_rr_client,
-                cluster_id,
-                peer_is_rr_client,
-            ) {
-                if rib_out.get_evpn(key).is_some() {
-                    evpn_withdraw.push(*key);
-                }
-                continue;
-            }
-
-            // Export policy evaluation. Type 5 EVPN (IP Prefix per RFC 9136)
-            // carries a real IP prefix in its NLRI, so prefix-based policy
-            // clauses can match against it directly. Types 1-4 don't carry
-            // a prefix in any meaningful sense (they identify Ethernet
-            // Segments, MAC addresses, or multicast tags), so prefix-based
-            // predicates do not match them. Operators who need to filter
-            // Types 1-4 should use RT- or community-based clauses.
-            let policy_prefix = match &best.route {
-                rustbgpd_wire::EvpnRoute::IpPrefix(t5) => match t5.prefix {
-                    rustbgpd_wire::EvpnIpPrefixValue::V4(p) => Some(Prefix::V4(p)),
-                    rustbgpd_wire::EvpnIpPrefixValue::V6(p) => Some(Prefix::V6(p)),
-                },
-                _ => None,
-            };
-            let aspath_str = if needs_as_path_string {
-                best.as_path()
-                    .map_or_else(String::new, rustbgpd_wire::AsPath::to_aspath_string)
-            } else {
-                String::new()
-            };
-            let aspath_len = best.as_path().map_or(0, rustbgpd_wire::AsPath::len);
-            let ctx = RouteContext {
-                prefix: policy_prefix,
-                next_hop: Some(best.next_hop),
-                extended_communities: best.extended_communities(),
-                communities: best.communities(),
-                large_communities: best.large_communities(),
-                as_path_str: &aspath_str,
-                as_path_len: aspath_len,
-                validation_state: rustbgpd_wire::RpkiValidation::NotFound,
-                aspa_state: rustbgpd_wire::AspaValidation::Unknown,
-                peer_address: Some(target_peer),
-                peer_asn: target_peer_asn,
-                peer_group: target_peer_group,
-                route_type: Some(route_type(best.origin_type)),
-                evpn_route_type: Some(best.route_type()),
-                local_pref: best.local_pref_attr(),
-                med: best.med_attr(),
-            };
-            let (result, evaluation) =
-                rustbgpd_policy::evaluate_chain_with_attribution(export_pol, &ctx);
-            record_export_policy_eval(metrics, policy_stats, target_peer_label, &evaluation);
-            if result.action != rustbgpd_policy::PolicyAction::Permit {
-                if rib_out.get_evpn(key).is_some() {
-                    evpn_withdraw.push(*key);
-                }
-                continue;
-            }
-
-            // Apply export modifications (community add/remove, RT add/remove,
-            // next-hop). Skip the deep attribute clone when nothing changes.
-            let mut modified = best.clone();
-            if !result.modifications.is_empty() {
-                let nh = rustbgpd_policy::apply_modifications(
-                    std::sync::Arc::make_mut(&mut modified.attributes),
-                    &result.modifications,
-                );
-                if let Some(rustbgpd_policy::NextHopAction::Specific(addr)) = nh {
-                    modified.next_hop = addr;
-                }
-            }
-            // Skip the announce if rib_out already holds the same route —
-            // the dirty-resync path in particular re-evaluates every
-            // advertised key and would otherwise re-emit no-op announces
-            // every cycle on a large fabric.
-            //
-            // `force` mode bypasses this so currently-advertised EVPN
-            // routes re-emit even when the RIB-level attribute set is
-            // unchanged — used for outbound-attribute toggles like the
-            // RFC 8326 GShut community attach where the wire change is
-            // applied later in transport.
-            if !force
-                && rib_out
-                    .get_evpn(key)
-                    .is_some_and(|existing| evpn_routes_equal(existing, &modified))
-            {
-                continue;
-            }
-            evpn_announce.push(modified);
-        }
-    }
-
-    /// Stage BGP-LS announces and withdrawals for a set of affected opaque keys.
-    ///
-    /// BGP-LS remains opaque to the RIB: selection and reflection use the
-    /// BGP attributes and the route key, not semantic TLV parsing. Policy gets
-    /// the same placeholder-prefix treatment as non-Type-5 EVPN; operators can
-    /// match communities, large communities, `AS_PATH`, peer, and route type, but
-    /// BGP-LS-specific predicates are intentionally deferred.
-    #[expect(
-        clippy::too_many_arguments,
-        clippy::too_many_lines,
-        reason = "BGP-LS staging mirrors EVPN distribution context for RR/export parity"
-    )]
-    pub(super) fn stage_bgpls_routes(
-        loc_rib: &LocRib,
-        rib_out: &AdjRibOut,
-        peer_is_rr_client: &HashMap<IpAddr, bool>,
-        keys: &HashSet<crate::route::BgpLsRouteKey>,
-        target_peer: IpAddr,
-        target_peer_asn: Option<u32>,
-        target_peer_group: Option<&str>,
-        target_is_ebgp: bool,
-        target_is_rr_client: bool,
-        cluster_id: Option<Ipv4Addr>,
-        sendable: Option<&Vec<(Afi, Safi)>>,
-        export_pol: Option<&PolicyChain>,
-        metrics: &BgpMetrics,
-        policy_stats: &mut NeighborPolicyStats,
-        target_peer_label: &str,
-        bgpls_announce: &mut Vec<crate::route::BgpLsRibRoute>,
-        bgpls_withdraw: &mut Vec<crate::route::BgpLsRouteKey>,
-        force: bool,
-    ) {
-        let needs_as_path_string = export_pol.is_some_and(PolicyChain::requires_as_path_string);
-        for key in keys {
-            let family = key.family.to_afi_safi();
-            if !sendable.is_some_and(|f| f.contains(&family)) {
-                if rib_out.get_bgpls(key).is_some() {
-                    bgpls_withdraw.push(key.clone());
-                }
-                continue;
-            }
-
-            let Some(best) = loc_rib.get_bgpls(key) else {
-                if rib_out.get_bgpls(key).is_some() {
-                    bgpls_withdraw.push(key.clone());
-                }
-                continue;
-            };
-
-            if best.peer == target_peer {
-                if rib_out.get_bgpls(key).is_some() {
-                    bgpls_withdraw.push(key.clone());
-                }
-                continue;
-            }
-
-            let probe = crate::route::Route {
-                prefix: Prefix::V4(rustbgpd_wire::Ipv4Prefix::new(Ipv4Addr::UNSPECIFIED, 0)),
-                next_hop: best.next_hop,
-                link_local_next_hop: None,
-                next_hop_scope: None,
-                peer: best.peer,
-                attributes: std::sync::Arc::new(vec![]),
-                received_at: best.received_at,
-                origin_type: best.origin_type,
-                peer_router_id: best.peer_router_id,
-                is_stale: false,
-                is_llgr_stale: false,
-                path_id: 0,
-                validation_state: rustbgpd_wire::RpkiValidation::NotFound,
-                aspa_state: rustbgpd_wire::AspaValidation::Unknown,
-                aspa_context: rustbgpd_wire::AspaValidationContext::default(),
-            };
-            if should_suppress_ibgp_inner(
-                &probe,
-                target_is_ebgp,
-                target_is_rr_client,
-                cluster_id,
-                peer_is_rr_client,
-            ) {
-                if rib_out.get_bgpls(key).is_some() {
-                    bgpls_withdraw.push(key.clone());
-                }
-                continue;
-            }
-
-            let aspath_str = if needs_as_path_string {
-                best.as_path()
-                    .map_or_else(String::new, rustbgpd_wire::AsPath::to_aspath_string)
-            } else {
-                String::new()
-            };
-            let aspath_len = best.as_path().map_or(0, rustbgpd_wire::AsPath::len);
-            let ctx = RouteContext {
-                prefix: None,
-                next_hop: Some(best.next_hop),
-                extended_communities: best.extended_communities(),
-                communities: best.communities(),
-                large_communities: best.large_communities(),
-                as_path_str: &aspath_str,
-                as_path_len: aspath_len,
-                validation_state: rustbgpd_wire::RpkiValidation::NotFound,
-                aspa_state: rustbgpd_wire::AspaValidation::Unknown,
-                peer_address: Some(target_peer),
-                peer_asn: target_peer_asn,
-                peer_group: target_peer_group,
-                route_type: Some(route_type(best.origin_type)),
-                evpn_route_type: None,
-                local_pref: best.local_pref_attr(),
-                med: best.med_attr(),
-            };
-            let (result, evaluation) =
-                rustbgpd_policy::evaluate_chain_with_attribution(export_pol, &ctx);
-            record_export_policy_eval(metrics, policy_stats, target_peer_label, &evaluation);
-            if result.action != rustbgpd_policy::PolicyAction::Permit {
-                if rib_out.get_bgpls(key).is_some() {
-                    bgpls_withdraw.push(key.clone());
-                }
-                continue;
-            }
-
-            let mut modified = best.clone();
-            if !result.modifications.is_empty() {
-                let nh = rustbgpd_policy::apply_modifications(
-                    std::sync::Arc::make_mut(&mut modified.attributes),
-                    &result.modifications,
-                );
-                if let Some(rustbgpd_policy::NextHopAction::Specific(addr)) = nh {
-                    modified.next_hop = addr;
-                }
-            }
-            modified.path_id = 0;
-
-            if !force
-                && rib_out
-                    .get_bgpls(key)
-                    .is_some_and(|existing| bgpls_routes_equal(existing, &modified))
-            {
-                continue;
-            }
-            bgpls_announce.push(modified);
-        }
-    }
-
-    /// Resolve the RFC 4684 VPN outbound filter for `peer`: `Some` iff the
-    /// peer negotiated `(IPv4, RtConstrain)`; a somehow-absent membership
-    /// entry resolves to the strict empty filter (advertise nothing) rather
-    /// than fail-open. Returns an owned clone because every caller holds a
-    /// `&mut` Adj-RIB-Out borrow through `self` while staging (the ORF
-    /// precedent) — memberships are small, present only for RTC peers.
-    pub(super) fn rtc_vpn_filter(
-        &self,
-        peer: IpAddr,
-        sendable: Option<&Vec<(Afi, Safi)>>,
-    ) -> Option<super::RtcMembership> {
-        if sendable.is_some_and(|f| f.contains(&crate::route::RtcRibRouteKey::afi_safi())) {
-            Some(
-                self.peer_rt_membership
-                    .get(&peer)
-                    .cloned()
-                    .unwrap_or_default(),
-            )
-        } else {
-            None
-        }
-    }
-
-    /// Stage VPNv4/VPNv6 announces and withdrawals for a set of affected keys.
-    ///
-    /// ADR-0077 §6 guardrail: the next-hop, MPLS label stack, and Route
-    /// Distinguisher pass through reflection unchanged — the staged route
-    /// carries the original `VpnNlri` verbatim and transport never rewrites
-    /// the VPN next-hop. Export policy matches on the RD-scoped inner prefix
-    /// (honest, unlike the prefixless BGP-LS placeholder) plus communities,
-    /// large communities, `AS_PATH`, peer, and route type.
-    #[expect(
-        clippy::too_many_arguments,
-        clippy::too_many_lines,
-        reason = "VPN staging mirrors EVPN/BGP-LS distribution context for RR/export parity"
-    )]
-    pub(super) fn stage_vpn_routes(
-        loc_rib: &LocRib,
-        rib_out: &AdjRibOut,
-        peer_is_rr_client: &HashMap<IpAddr, bool>,
-        keys: &HashSet<crate::route::VpnRibRouteKey>,
-        target_peer: IpAddr,
-        target_peer_asn: Option<u32>,
-        target_peer_group: Option<&str>,
-        target_is_ebgp: bool,
-        target_is_rr_client: bool,
-        cluster_id: Option<Ipv4Addr>,
-        sendable: Option<&Vec<(Afi, Safi)>>,
-        rtc_filter: Option<&super::RtcMembership>,
-        export_pol: Option<&PolicyChain>,
-        metrics: &BgpMetrics,
-        policy_stats: &mut NeighborPolicyStats,
-        target_peer_label: &str,
-        vpn_announce: &mut Vec<crate::route::VpnRibRoute>,
-        vpn_withdraw: &mut Vec<crate::route::VpnRibRouteKey>,
-        force: bool,
-    ) {
-        let needs_as_path_string = export_pol.is_some_and(PolicyChain::requires_as_path_string);
-        for key in keys {
-            let family = key.afi_safi();
-            if !sendable.is_some_and(|f| f.contains(&family)) {
-                if rib_out.get_vpn(key).is_some() {
-                    vpn_withdraw.push(key.clone());
-                }
-                continue;
-            }
-
-            let Some(best) = loc_rib.get_vpn(key) else {
-                if rib_out.get_vpn(key).is_some() {
-                    vpn_withdraw.push(key.clone());
-                }
-                continue;
-            };
-
-            // RFC 4684 outbound gate: a peer that negotiated RT-Constrain
-            // only receives VPN routes whose Route Targets fall inside its
-            // advertised membership (`None` = SAFI 132 not negotiated ⇒
-            // unfiltered). Membership is per-peer, so the one gate covers
-            // both VPNv4 and VPNv6 keys. Miss ⇒ withdraw-if-present,
-            // exactly the sendable-family gate shape above.
-            if let Some(membership) = rtc_filter
-                && !membership.matches_any(best.extended_communities())
-            {
-                if rib_out.get_vpn(key).is_some() {
-                    vpn_withdraw.push(key.clone());
-                }
-                continue;
-            }
-
-            if best.peer == target_peer {
-                if rib_out.get_vpn(key).is_some() {
-                    vpn_withdraw.push(key.clone());
-                }
-                continue;
-            }
-
-            let probe = crate::route::Route {
-                prefix: best.inner_prefix(),
-                next_hop: best.next_hop,
-                link_local_next_hop: None,
-                next_hop_scope: None,
-                peer: best.peer,
-                attributes: std::sync::Arc::new(vec![]),
-                received_at: best.received_at,
-                origin_type: best.origin_type,
-                peer_router_id: best.peer_router_id,
-                is_stale: false,
-                is_llgr_stale: false,
-                path_id: 0,
-                validation_state: rustbgpd_wire::RpkiValidation::NotFound,
-                aspa_state: rustbgpd_wire::AspaValidation::Unknown,
-                aspa_context: rustbgpd_wire::AspaValidationContext::default(),
-            };
-            if should_suppress_ibgp_inner(
-                &probe,
-                target_is_ebgp,
-                target_is_rr_client,
-                cluster_id,
-                peer_is_rr_client,
-            ) {
-                if rib_out.get_vpn(key).is_some() {
-                    vpn_withdraw.push(key.clone());
-                }
-                continue;
-            }
-
-            let aspath_str = if needs_as_path_string {
-                best.as_path()
-                    .map_or_else(String::new, rustbgpd_wire::AsPath::to_aspath_string)
-            } else {
-                String::new()
-            };
-            let aspath_len = best.as_path().map_or(0, rustbgpd_wire::AsPath::len);
-            let ctx = RouteContext {
-                prefix: Some(best.inner_prefix()),
-                next_hop: Some(best.next_hop),
-                extended_communities: best.extended_communities(),
-                communities: best.communities(),
-                large_communities: best.large_communities(),
-                as_path_str: &aspath_str,
-                as_path_len: aspath_len,
-                validation_state: rustbgpd_wire::RpkiValidation::NotFound,
-                aspa_state: rustbgpd_wire::AspaValidation::Unknown,
-                peer_address: Some(target_peer),
-                peer_asn: target_peer_asn,
-                peer_group: target_peer_group,
-                route_type: Some(route_type(best.origin_type)),
-                evpn_route_type: None,
-                local_pref: best.local_pref_attr(),
-                med: best.med_attr(),
-            };
-            let (result, evaluation) =
-                rustbgpd_policy::evaluate_chain_with_attribution(export_pol, &ctx);
-            record_export_policy_eval(metrics, policy_stats, target_peer_label, &evaluation);
-            if result.action != rustbgpd_policy::PolicyAction::Permit {
-                if rib_out.get_vpn(key).is_some() {
-                    vpn_withdraw.push(key.clone());
-                }
-                continue;
-            }
-
-            let mut modified = best.clone();
-            if !result.modifications.is_empty() {
-                let nh = rustbgpd_policy::apply_modifications(
-                    std::sync::Arc::make_mut(&mut modified.attributes),
-                    &result.modifications,
-                );
-                if let Some(rustbgpd_policy::NextHopAction::Specific(addr)) = nh {
-                    modified.next_hop = addr;
-                }
-            }
-            modified.path_id = 0;
-
-            if !force
-                && rib_out
-                    .get_vpn(key)
-                    .is_some_and(|existing| vpn_routes_equal(existing, &modified))
-            {
-                continue;
-            }
-            vpn_announce.push(modified);
-        }
-    }
-
-    /// Stage RT-Constrain announces and withdrawals for a set of affected
-    /// keys. Mirrors [`Self::stage_vpn_routes`] minus labels; like BGP-LS,
-    /// the export-policy context carries no prefix (`prefix: None`) — an RT
-    /// membership NLRI has no IP prefix to match on.
-    #[expect(
-        clippy::too_many_arguments,
-        clippy::too_many_lines,
-        reason = "RTC staging mirrors VPN/BGP-LS distribution context for RR/export parity"
-    )]
-    pub(super) fn stage_rtc_routes(
-        loc_rib: &LocRib,
-        rib_out: &AdjRibOut,
-        peer_is_rr_client: &HashMap<IpAddr, bool>,
-        keys: &HashSet<crate::route::RtcRibRouteKey>,
-        target_peer: IpAddr,
-        target_peer_asn: Option<u32>,
-        target_peer_group: Option<&str>,
-        target_is_ebgp: bool,
-        target_is_rr_client: bool,
-        cluster_id: Option<Ipv4Addr>,
-        sendable: Option<&Vec<(Afi, Safi)>>,
-        export_pol: Option<&PolicyChain>,
-        metrics: &BgpMetrics,
-        policy_stats: &mut NeighborPolicyStats,
-        target_peer_label: &str,
-        rtc_announce: &mut Vec<crate::route::RtcRibRoute>,
-        rtc_withdraw: &mut Vec<crate::route::RtcRibRouteKey>,
-        force: bool,
-    ) {
-        let needs_as_path_string = export_pol.is_some_and(PolicyChain::requires_as_path_string);
-        let family = crate::route::RtcRibRouteKey::afi_safi();
-        let family_sendable = sendable.is_some_and(|f| f.contains(&family));
-        for key in keys {
-            if !family_sendable {
-                if rib_out.get_rtc(key).is_some() {
-                    rtc_withdraw.push(key.clone());
-                }
-                continue;
-            }
-
-            let Some(best) = loc_rib.get_rtc(key) else {
-                if rib_out.get_rtc(key).is_some() {
-                    rtc_withdraw.push(key.clone());
-                }
-                continue;
-            };
-
-            if best.peer == target_peer {
-                if rib_out.get_rtc(key).is_some() {
-                    rtc_withdraw.push(key.clone());
-                }
-                continue;
-            }
-
-            let probe = crate::route::Route {
-                prefix: Prefix::V4(rustbgpd_wire::Ipv4Prefix::new(Ipv4Addr::UNSPECIFIED, 0)),
-                next_hop: best.next_hop,
-                link_local_next_hop: None,
-                next_hop_scope: None,
-                peer: best.peer,
-                attributes: std::sync::Arc::new(vec![]),
-                received_at: best.received_at,
-                origin_type: best.origin_type,
-                peer_router_id: best.peer_router_id,
-                is_stale: false,
-                is_llgr_stale: false,
-                path_id: 0,
-                validation_state: rustbgpd_wire::RpkiValidation::NotFound,
-                aspa_state: rustbgpd_wire::AspaValidation::Unknown,
-                aspa_context: rustbgpd_wire::AspaValidationContext::default(),
-            };
-            if should_suppress_ibgp_inner(
-                &probe,
-                target_is_ebgp,
-                target_is_rr_client,
-                cluster_id,
-                peer_is_rr_client,
-            ) {
-                if rib_out.get_rtc(key).is_some() {
-                    rtc_withdraw.push(key.clone());
-                }
-                continue;
-            }
-
-            let aspath_str = if needs_as_path_string {
-                best.as_path()
-                    .map_or_else(String::new, rustbgpd_wire::AsPath::to_aspath_string)
-            } else {
-                String::new()
-            };
-            let aspath_len = best.as_path().map_or(0, rustbgpd_wire::AsPath::len);
-            let ctx = RouteContext {
-                prefix: None,
-                next_hop: Some(best.next_hop),
-                extended_communities: best.extended_communities(),
-                communities: best.communities(),
-                large_communities: best.large_communities(),
-                as_path_str: &aspath_str,
-                as_path_len: aspath_len,
-                validation_state: rustbgpd_wire::RpkiValidation::NotFound,
-                aspa_state: rustbgpd_wire::AspaValidation::Unknown,
-                peer_address: Some(target_peer),
-                peer_asn: target_peer_asn,
-                peer_group: target_peer_group,
-                route_type: Some(route_type(best.origin_type)),
-                evpn_route_type: None,
-                local_pref: best.local_pref_attr(),
-                med: best.med_attr(),
-            };
-            let (result, evaluation) =
-                rustbgpd_policy::evaluate_chain_with_attribution(export_pol, &ctx);
-            record_export_policy_eval(metrics, policy_stats, target_peer_label, &evaluation);
-            if result.action != rustbgpd_policy::PolicyAction::Permit {
-                if rib_out.get_rtc(key).is_some() {
-                    rtc_withdraw.push(key.clone());
-                }
-                continue;
-            }
-
-            let mut modified = best.clone();
-            if !result.modifications.is_empty() {
-                let nh = rustbgpd_policy::apply_modifications(
-                    std::sync::Arc::make_mut(&mut modified.attributes),
-                    &result.modifications,
-                );
-                if let Some(rustbgpd_policy::NextHopAction::Specific(addr)) = nh {
-                    modified.next_hop = addr;
-                }
-            }
-            modified.path_id = 0;
-
-            if !force
-                && rib_out
-                    .get_rtc(key)
-                    .is_some_and(|existing| rtc_routes_equal(existing, &modified))
-            {
-                continue;
-            }
-            rtc_announce.push(modified);
-        }
-    }
-
-    /// Distribute Loc-RIB changes to all registered outbound peers.
-    ///
-    /// For clean peers, only `changed_prefixes` are evaluated. Dirty peers
-    /// (those that failed a previous `try_send()`) get a full export resync:
-    /// all Loc-RIB and `AdjRibOut` prefixes are diffed to bring the peer's
-    /// view back in sync. `AdjRibOut` is only committed after a successful
-    /// channel send; on failure the peer stays dirty for retry via the
-    /// resync timer.
-    ///
-    /// Routes are filtered by `sendable_families` (set at `PeerUp` time)
-    /// so that Adj-RIB-Out only contains routes the transport can actually
-    /// serialize for this peer. The transport retains `is_family_negotiated`
-    /// as a safety net.
-    #[expect(
-        clippy::too_many_lines,
-        reason = "distribution loop coordinates dirty peers, forced resync, and all families"
-    )]
-    pub(super) fn distribute_changes(
-        &mut self,
-        best_changed: &HashSet<Prefix>,
-        all_affected: &HashSet<Prefix>,
-    ) {
-        if best_changed.is_empty()
-            && all_affected.is_empty()
-            && self.dirty_peers.is_empty()
-            && self.force_outbound_peers.is_empty()
-        {
-            return;
-        }
-
-        let peers: Vec<IpAddr> = self.outbound_peers.keys().copied().collect();
-        for peer in peers {
-            // For dirty peers, compute full prefix set from Loc-RIB + AdjRibOut
-            let is_dirty = self.dirty_peers.contains(&peer);
-            // `is_force` peers are dirty-equivalent for prefix enumeration AND
-            // bypass the AdjRibOut equality suppression in the per-prefix
-            // helpers — see `RibUpdate::RefreshPeerOutbound` rationale on
-            // `force_outbound_peers`.
-            let is_force = self.force_outbound_peers.contains(&peer);
-            let resync = is_dirty || is_force;
-            let effective_prefixes: HashSet<Prefix> = if resync {
-                let mut all: HashSet<Prefix> = self.loc_rib.iter().map(|r| r.prefix).collect();
-                // For multi-path dirty resync, also include all Adj-RIB-In prefixes
-                if self.peer_has_any_add_path_send(peer) {
-                    for rib in self.ribs.values() {
-                        all.extend(rib.iter().map(|r| r.prefix));
-                    }
-                }
-                if let Some(rib_out) = self.adj_ribs_out.get(&peer) {
-                    all.extend(rib_out.iter().map(|r| r.prefix));
-                }
-                all
-            } else {
-                let mut prefixes = best_changed.clone();
-                // An RFC 9107 ORR peer selects from the per-target
-                // candidate set, not the Loc-RIB best — a candidate
-                // change that leaves the Loc-RIB best untouched can
-                // still flip the vantage best, so ORR peers stage every
-                // affected prefix (same reasoning as Add-Path send).
-                let peer_has_resolved_orr = self
-                    .peer_orr_vantage
-                    .get(&peer)
-                    .is_some_and(|vantage| self.orr.spf.contains_key(vantage));
-                for prefix in all_affected {
-                    if peer_has_resolved_orr || self.add_path_send_max_for_prefix(peer, prefix) > 0
-                    {
-                        prefixes.insert(*prefix);
-                    }
-                }
-                prefixes
-            };
-            let effective_flowspec_rules: HashSet<FlowSpecRule> = if resync {
-                let mut all: HashSet<FlowSpecRule> = self
-                    .loc_rib
-                    .iter_flowspec()
-                    .map(|route| route.rule.clone())
-                    .collect();
-                if let Some(rib_out) = self.adj_ribs_out.get(&peer) {
-                    all.extend(rib_out.iter_flowspec().map(|route| route.rule.clone()));
-                }
-                all
-            } else {
-                HashSet::new()
-            };
-
-            let effective_evpn_keys: HashSet<rustbgpd_wire::EvpnRouteKey> = if resync {
-                let mut all: HashSet<rustbgpd_wire::EvpnRouteKey> = self
-                    .loc_rib
-                    .iter_evpn()
-                    .map(crate::route::EvpnRibRoute::key)
-                    .collect();
-                if let Some(rib_out) = self.adj_ribs_out.get(&peer) {
-                    all.extend(rib_out.iter_evpn().map(crate::route::EvpnRibRoute::key));
-                }
-                all
-            } else {
-                HashSet::new()
-            };
-
-            let effective_bgpls_keys: HashSet<crate::route::BgpLsRouteKey> = if resync {
-                let mut all: HashSet<crate::route::BgpLsRouteKey> = self
-                    .loc_rib
-                    .iter_bgpls()
-                    .map(crate::route::BgpLsRibRoute::key)
-                    .collect();
-                if let Some(rib_out) = self.adj_ribs_out.get(&peer) {
-                    all.extend(rib_out.iter_bgpls().map(crate::route::BgpLsRibRoute::key));
-                }
-                all
-            } else {
-                HashSet::new()
-            };
-
-            let effective_l3vpn_keys: HashSet<crate::route::VpnRibRouteKey> = if resync {
-                let mut all: HashSet<crate::route::VpnRibRouteKey> = self
-                    .loc_rib
-                    .iter_vpn()
-                    .map(crate::route::VpnRibRoute::key)
-                    .collect();
-                if let Some(rib_out) = self.adj_ribs_out.get(&peer) {
-                    all.extend(rib_out.iter_vpn().map(crate::route::VpnRibRoute::key));
-                }
-                all
-            } else {
-                HashSet::new()
-            };
-
-            let effective_rtc_keys: HashSet<crate::route::RtcRibRouteKey> = if resync {
-                let mut all: HashSet<crate::route::RtcRibRouteKey> = self
-                    .loc_rib
-                    .iter_rtc()
-                    .map(crate::route::RtcRibRoute::key)
-                    .collect();
-                if let Some(rib_out) = self.adj_ribs_out.get(&peer) {
-                    all.extend(rib_out.iter_rtc().map(crate::route::RtcRibRoute::key));
-                }
-                all
-            } else {
-                HashSet::new()
-            };
-
-            if effective_prefixes.is_empty()
-                && effective_flowspec_rules.is_empty()
-                && effective_evpn_keys.is_empty()
-                && effective_bgpls_keys.is_empty()
-                && effective_l3vpn_keys.is_empty()
-                && effective_rtc_keys.is_empty()
-            {
-                self.clear_policy_filtered_routes_for_peer(peer);
-                // Resync flags must clear here too — otherwise a
-                // force-only refresh on a peer with no exportable
-                // routes would leave `force_outbound_peers` populated,
-                // and the next unrelated dirty resync (or another
-                // RefreshPeerOutbound for a different reason) would
-                // accidentally inherit the bypass-equality-suppression
-                // semantics. Same shape for `dirty_peers` for symmetry.
-                if is_dirty {
-                    self.dirty_peers.remove(&peer);
-                }
-                if is_force {
-                    self.force_outbound_peers.remove(&peer);
-                }
-                continue;
-            }
-
-            let mut announce = Vec::new();
-            let mut withdraw = Vec::new();
-            let mut nh_override_flags: Vec<Option<rustbgpd_policy::NextHopAction>> = Vec::new();
-            let mut fs_announce = Vec::new();
-            let mut fs_withdraw = Vec::new();
-            let mut evpn_announce = Vec::new();
-            let mut evpn_withdraw = Vec::new();
-            let mut bgpls_announce = Vec::new();
-            let mut bgpls_withdraw = Vec::new();
-            let mut vpn_announce = Vec::new();
-            let mut vpn_withdraw = Vec::new();
-            let mut rtc_announce = Vec::new();
-            let mut rtc_withdraw = Vec::new();
-            let mut current_policy_filtered_routes: HashSet<PolicyFilteredRouteKey> =
-                HashSet::new();
-
-            // Resolve export policy, sendable families, and RR state before
-            // borrowing rib_out (which holds a &mut to self.adj_ribs_out).
-            let export_pol = self.export_policy_for(peer).cloned();
-            let sendable = self.peer_sendable_families.get(&peer).cloned();
-            let target_is_ebgp = self.peer_is_ebgp.get(&peer).copied().unwrap_or(true);
-            let target_is_rr_client = self.peer_is_rr_client.get(&peer).copied().unwrap_or(false);
-            let target_peer_asn = self.peer_asn.get(&peer).copied();
-            let target_peer_group = self.peer_group.get(&peer).map(String::as_str);
-            let cluster_id = self.cluster_id;
-            let peer_add_path_send_max =
-                self.peer_add_path_send_max.get(&peer).copied().unwrap_or(0);
-            let peer_add_path_send_families = self
-                .peer_add_path_send_families
-                .get(&peer)
-                .cloned()
-                .unwrap_or_default();
-            // ORF state, resolved before the &mut rib_out borrow below. Cloning
-            // is cheap: ORF filters are small and present only for peers that
-            // negotiated ORF (None for everyone else).
-            let orf_filters = self.peer_orf_filters.get(&peer).cloned();
-            // RFC 9107 ORR: a peer bound to a vantage that resolved this
-            // pass takes the per-vantage best below; an unresolved
-            // vantage silently falls back to the standard single-best
-            // (the status is surfaced by `rbgp orr`).
-            let orr_ctx = self
-                .peer_orr_vantage
-                .get(&peer)
-                .and_then(|vantage| self.orr.spf.get(vantage))
-                .map(|spf| (&self.orr.topology, spf));
-            let rtc_filter = self.rtc_vpn_filter(peer, sendable.as_ref());
-            let orf_gated = self
-                .peer_orf_pending
-                .get(&peer)
-                .cloned()
-                .unwrap_or_default();
-            let loc_rib = &self.loc_rib;
-            let loc_rib_len = loc_rib.len();
-            let target_peer_label = peer.to_string();
-            let metrics = self.metrics.clone();
-            let policy_stats = self.export_policy_stats.entry(peer).or_default();
-
-            let rib_out = self
-                .adj_ribs_out
-                .entry(peer)
-                .or_insert_with(|| AdjRibOut::with_capacity(peer, loc_rib_len));
-
-            // Stage: compute delta without mutating AdjRibOut
-            for prefix in &effective_prefixes {
-                let family = prefix_family(prefix);
-                // RFC 5291 §6 gate: suppress this family's advertisement (incl.
-                // ongoing churn) until the peer's first ROUTE-REFRESH lifts it.
-                if orf_gated.contains(&family) {
-                    continue;
-                }
-                let orf = orf_filters.as_ref().and_then(|m| m.get(&family));
-                let prefix_send_max = if peer_add_path_send_max > 0
-                    && peer_add_path_send_families.contains(&family)
-                {
-                    peer_add_path_send_max
-                } else {
-                    0
-                };
-                if prefix_send_max > 0 {
-                    // Multi-path: collect all candidates, filter, sort, diff
-                    let mut policy_filtered = Vec::new();
-                    Self::distribute_multipath_prefix(
-                        &self.ribs,
-                        rib_out,
-                        &self.peer_is_rr_client,
-                        prefix,
-                        peer,
-                        target_peer_asn,
-                        target_peer_group,
-                        prefix_send_max,
-                        target_is_ebgp,
-                        target_is_rr_client,
-                        cluster_id,
-                        sendable.as_ref(),
-                        export_pol.as_ref(),
-                        orf,
-                        orr_ctx,
-                        &metrics,
-                        policy_stats,
-                        &target_peer_label,
-                        &mut announce,
-                        &mut withdraw,
-                        &mut nh_override_flags,
-                        &mut policy_filtered,
-                        is_force,
-                    );
-                    current_policy_filtered_routes.extend(policy_filtered);
-                } else if let Some((orr_topology, orr_spf)) = orr_ctx {
-                    // ORR peer with a resolved vantage: per-vantage best.
-                    let mut policy_filtered = Vec::new();
-                    Self::distribute_orr_best_prefix(
-                        &self.ribs,
-                        rib_out,
-                        &self.peer_is_rr_client,
-                        orr_topology,
-                        orr_spf,
-                        prefix,
-                        peer,
-                        target_peer_asn,
-                        target_peer_group,
-                        target_is_ebgp,
-                        target_is_rr_client,
-                        cluster_id,
-                        sendable.as_ref(),
-                        export_pol.as_ref(),
-                        orf,
-                        &metrics,
-                        policy_stats,
-                        &target_peer_label,
-                        &mut announce,
-                        &mut withdraw,
-                        &mut nh_override_flags,
-                        &mut policy_filtered,
-                        is_force,
-                    );
-                    current_policy_filtered_routes.extend(policy_filtered);
-                } else {
-                    let mut policy_filtered = Vec::new();
-                    Self::distribute_single_best_prefix(
-                        loc_rib,
-                        rib_out,
-                        &self.peer_is_rr_client,
-                        prefix,
-                        peer,
-                        target_peer_asn,
-                        target_peer_group,
-                        target_is_ebgp,
-                        target_is_rr_client,
-                        cluster_id,
-                        sendable.as_ref(),
-                        export_pol.as_ref(),
-                        orf,
-                        &metrics,
-                        policy_stats,
-                        &target_peer_label,
-                        &mut announce,
-                        &mut withdraw,
-                        &mut nh_override_flags,
-                        &mut policy_filtered,
-                        is_force,
-                    );
-                    current_policy_filtered_routes.extend(policy_filtered);
-                }
-            }
-
-            if resync && !effective_flowspec_rules.is_empty() {
-                Self::stage_flowspec_rules(
-                    loc_rib,
-                    rib_out,
-                    &self.peer_is_rr_client,
-                    &effective_flowspec_rules,
-                    peer,
-                    target_peer_asn,
-                    target_peer_group,
-                    target_is_ebgp,
-                    target_is_rr_client,
-                    cluster_id,
-                    sendable.as_ref(),
-                    export_pol.as_ref(),
-                    &metrics,
-                    policy_stats,
-                    &target_peer_label,
-                    &mut fs_announce,
-                    &mut fs_withdraw,
-                );
-            }
-
-            if resync && !effective_evpn_keys.is_empty() {
-                Self::stage_evpn_routes(
-                    loc_rib,
-                    rib_out,
-                    &self.peer_is_rr_client,
-                    &effective_evpn_keys,
-                    peer,
-                    target_peer_asn,
-                    target_peer_group,
-                    target_is_ebgp,
-                    target_is_rr_client,
-                    cluster_id,
-                    sendable.as_ref(),
-                    export_pol.as_ref(),
-                    &metrics,
-                    policy_stats,
-                    &target_peer_label,
-                    &mut evpn_announce,
-                    &mut evpn_withdraw,
-                    is_force,
-                );
-            }
-
-            if resync && !effective_bgpls_keys.is_empty() {
-                Self::stage_bgpls_routes(
-                    loc_rib,
-                    rib_out,
-                    &self.peer_is_rr_client,
-                    &effective_bgpls_keys,
-                    peer,
-                    target_peer_asn,
-                    target_peer_group,
-                    target_is_ebgp,
-                    target_is_rr_client,
-                    cluster_id,
-                    sendable.as_ref(),
-                    export_pol.as_ref(),
-                    &metrics,
-                    policy_stats,
-                    &target_peer_label,
-                    &mut bgpls_announce,
-                    &mut bgpls_withdraw,
-                    is_force,
-                );
-            }
-
-            if resync && !effective_l3vpn_keys.is_empty() {
-                Self::stage_vpn_routes(
-                    loc_rib,
-                    rib_out,
-                    &self.peer_is_rr_client,
-                    &effective_l3vpn_keys,
-                    peer,
-                    target_peer_asn,
-                    target_peer_group,
-                    target_is_ebgp,
-                    target_is_rr_client,
-                    cluster_id,
-                    sendable.as_ref(),
-                    rtc_filter.as_ref(),
-                    export_pol.as_ref(),
-                    &metrics,
-                    policy_stats,
-                    &target_peer_label,
-                    &mut vpn_announce,
-                    &mut vpn_withdraw,
-                    is_force,
-                );
-            }
-
-            if resync && !effective_rtc_keys.is_empty() {
-                Self::stage_rtc_routes(
-                    loc_rib,
-                    rib_out,
-                    &self.peer_is_rr_client,
-                    &effective_rtc_keys,
-                    peer,
-                    target_peer_asn,
-                    target_peer_group,
-                    target_is_ebgp,
-                    target_is_rr_client,
-                    cluster_id,
-                    sendable.as_ref(),
-                    export_pol.as_ref(),
-                    &metrics,
-                    policy_stats,
-                    &target_peer_label,
-                    &mut rtc_announce,
-                    &mut rtc_withdraw,
-                    is_force,
-                );
-            }
-
-            if !announce.is_empty()
-                || !withdraw.is_empty()
-                || !fs_announce.is_empty()
-                || !fs_withdraw.is_empty()
-                || !evpn_announce.is_empty()
-                || !evpn_withdraw.is_empty()
-                || !bgpls_announce.is_empty()
-                || !bgpls_withdraw.is_empty()
-                || !vpn_announce.is_empty()
-                || !vpn_withdraw.is_empty()
-                || !rtc_announce.is_empty()
-                || !rtc_withdraw.is_empty()
-            {
-                // If a prior initial dump / route-refresh EoR was deferred,
-                // piggyback it on the successful dirty resync update so it
-                // can't be starved behind the resync message on a small queue.
-                // EoR piggyback only attaches to *dirty* resyncs (the
-                // dump-deferral pattern). A force-only resync is a
-                // GShut-style outbound-attribute refresh and never
-                // carries pending EoR by definition.
-                let pending_eor = if is_dirty {
-                    self.pending_eor
-                        .get(&peer)
-                        .map(|families| families.iter().copied().collect())
-                        .unwrap_or_default()
-                } else {
-                    vec![]
-                };
-                let announced_count = announce.len();
-                let withdrawn_count = withdraw.len();
-                if self.try_send_and_commit_outbound_update(
-                    peer,
-                    nh_override_flags,
-                    announce,
-                    withdraw,
-                    pending_eor.clone(),
-                    vec![],
-                    fs_announce,
-                    fs_withdraw,
-                    evpn_announce,
-                    evpn_withdraw,
-                    bgpls_announce,
-                    bgpls_withdraw,
-                    vpn_announce,
-                    vpn_withdraw,
-                    rtc_announce,
-                    rtc_withdraw,
-                ) {
-                    self.update_policy_filtered_routes_for_prefixes(
-                        peer,
-                        &effective_prefixes,
-                        &current_policy_filtered_routes,
-                    );
-                    if resync {
-                        info!(
-                            %peer,
-                            announced = announced_count,
-                            withdrawn = withdrawn_count,
-                            dirty = is_dirty,
-                            force = is_force,
-                            "outbound routes resynced"
-                        );
-                        if is_dirty {
-                            self.dirty_peers.remove(&peer);
-                            if pending_eor.is_empty() {
-                                self.flush_pending_eor(peer);
-                            } else {
-                                self.pending_eor.remove(&peer);
-                            }
-                            self.retry_pending_refresh(peer);
-                        }
-                        // Force is one-shot: clear after a successful
-                        // emission so a subsequent unrelated dirty
-                        // resync doesn't re-bypass equality checks.
-                        if is_force {
-                            self.force_outbound_peers.remove(&peer);
-                        }
-                    }
-                } else {
-                    warn!(%peer, "outbound channel full or closed — marking dirty for resync");
-                    self.metrics.record_outbound_route_drop(&peer.to_string());
-                    self.dirty_peers.insert(peer);
-                }
-            } else {
-                self.update_policy_filtered_routes_for_prefixes(
-                    peer,
-                    &effective_prefixes,
-                    &current_policy_filtered_routes,
-                );
-                if resync {
-                    // Resync triggered but no diff — already in sync.
-                    debug!(%peer, "outbound routes unchanged after resync");
-                    if is_dirty {
-                        self.dirty_peers.remove(&peer);
-                        self.flush_pending_eor(peer);
-                        self.retry_pending_refresh(peer);
-                    }
-                    if is_force {
-                        self.force_outbound_peers.remove(&peer);
-                    }
-                }
-            }
-        }
-    }
-
-    /// Recompute `FlowSpec` Loc-RIB best routes for affected rules and
-    /// distribute changes to all outbound peers.
-    pub(super) fn recompute_and_distribute_flowspec(
-        &mut self,
-        affected: &HashSet<rustbgpd_wire::FlowSpecRule>,
-    ) {
-        use crate::route::FlowSpecRoute;
-
-        let mut changed_rules: HashSet<rustbgpd_wire::FlowSpecRule> = HashSet::new();
-
-        for rule in affected {
-            let candidates: Vec<&FlowSpecRoute> = self
-                .ribs
-                .values()
-                .flat_map(|rib| rib.iter_flowspec_rule(rule))
-                .collect();
-            let did_change = self
-                .loc_rib
-                .recompute_flowspec(rule.clone(), candidates.into_iter());
-            if did_change {
-                changed_rules.insert(rule.clone());
-            }
-        }
-
-        if changed_rules.is_empty() {
-            return;
-        }
-
-        self.metrics
-            .set_loc_rib_prefixes("flowspec", gauge_val(self.loc_rib.flowspec_len()));
-
-        // Distribute FlowSpec changes to outbound peers
-        let peers: Vec<IpAddr> = self.outbound_peers.keys().copied().collect();
-        for peer in peers {
-            let sendable = self.peer_sendable_families.get(&peer).cloned();
-            let has_fs = sendable.as_ref().is_some_and(|families| {
-                families.contains(&(rustbgpd_wire::Afi::Ipv4, rustbgpd_wire::Safi::FlowSpec))
-                    || families.contains(&(rustbgpd_wire::Afi::Ipv6, rustbgpd_wire::Safi::FlowSpec))
-            });
-            if !has_fs {
-                continue;
-            }
-
-            let target_is_ebgp = self.peer_is_ebgp.get(&peer).copied().unwrap_or(true);
-            let target_is_rr_client = self.peer_is_rr_client.get(&peer).copied().unwrap_or(false);
-            let target_peer_asn = self.peer_asn.get(&peer).copied();
-            let target_peer_group = self.peer_group.get(&peer).map(String::as_str);
-            let export_pol = self.export_policy_for(peer).cloned();
-            let target_peer_label = peer.to_string();
-            let metrics = self.metrics.clone();
-
-            let loc_rib_len = self.loc_rib.len();
-            let rib_out = self
-                .adj_ribs_out
-                .entry(peer)
-                .or_insert_with(|| crate::adj_rib_out::AdjRibOut::with_capacity(peer, loc_rib_len));
-            let policy_stats = self.export_policy_stats.entry(peer).or_default();
-
-            let mut fs_announce = Vec::new();
-            let mut fs_withdraw = Vec::new();
-            Self::stage_flowspec_rules(
-                &self.loc_rib,
-                rib_out,
-                &self.peer_is_rr_client,
-                &changed_rules,
-                peer,
-                target_peer_asn,
-                target_peer_group,
-                target_is_ebgp,
-                target_is_rr_client,
-                self.cluster_id,
-                sendable.as_ref(),
-                export_pol.as_ref(),
-                &metrics,
-                policy_stats,
-                &target_peer_label,
-                &mut fs_announce,
-                &mut fs_withdraw,
-            );
-
-            if (!fs_announce.is_empty() || !fs_withdraw.is_empty())
-                && !self.try_send_and_commit_outbound_update(
-                    peer,
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    fs_announce,
-                    fs_withdraw,
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                )
-            {
-                warn!(%peer, "outbound channel full — FlowSpec update deferred");
-                self.dirty_peers.insert(peer);
-            }
-        }
-    }
-
-    /// Recompute Loc-RIB best path and distribute changes for EVPN routes.
-    ///
-    /// Mirrors `recompute_and_distribute_flowspec`. For each affected key,
-    /// gather candidates from all peer Adj-RIB-Ins, run best-path, and if
-    /// the selection changed, stage announces/withdraws for each outbound
-    /// peer that negotiated the L2VPN/EVPN family.
-    #[expect(
-        clippy::too_many_lines,
-        reason = "EVPN recompute keeps best-path, events, and outbound staging consistent"
-    )]
-    pub(super) fn recompute_and_distribute_evpn(
-        &mut self,
-        affected: &HashSet<rustbgpd_wire::EvpnRouteKey>,
-    ) {
-        use crate::route::EvpnRibRoute;
-
-        let mut changed_keys: HashSet<rustbgpd_wire::EvpnRouteKey> = HashSet::new();
-        for key in affected {
-            let candidates: Vec<&EvpnRibRoute> = self
-                .ribs
-                .values()
-                .flat_map(|rib| rib.iter_evpn().filter(|r| r.key() == *key))
-                .collect();
-            // Capture old state BEFORE recompute_evpn so the event-type
-            // derivation below distinguishes Added / Withdrawn /
-            // BestChanged precisely. Inferring from the `did_change`
-            // bool alone collapses Added and BestChanged into one
-            // signal — Gate 7c subscribers (the daemon originator)
-            // need to know the difference to update their `last_seen
-            // _remote_seq` ratchets correctly.
-            let previous_best: Option<EvpnRibRoute> = self.loc_rib.get_evpn(key).cloned();
-            let did_change = self.loc_rib.recompute_evpn(*key, candidates.into_iter());
-            if did_change {
-                let new_best: Option<EvpnRibRoute> = self.loc_rib.get_evpn(key).cloned();
-                let event_type = match (&previous_best, &new_best) {
-                    (None, Some(_)) => crate::event::RouteEventType::Added,
-                    (Some(_), None) => crate::event::RouteEventType::Withdrawn,
-                    (Some(_), Some(_)) => crate::event::RouteEventType::BestChanged,
-                    // recompute returned `did_change=true` so at least
-                    // one side is `Some`. (None, None) shouldn't fire
-                    // — guard with debug_assert and skip silently.
-                    (None, None) => {
-                        debug_assert!(false, "recompute_evpn returned changed but both sides None");
-                        continue;
-                    }
-                };
-                let previous_peer = previous_best.as_ref().map(|r| r.peer);
-                let peer = new_best.as_ref().map(|r| r.peer);
-                self.publish_evpn_route_event(crate::event::EvpnRouteEvent {
-                    event_type,
-                    key: *key,
-                    best: new_best,
-                    previous_best,
-                    peer,
-                    previous_peer,
-                    timestamp: crate::event::unix_timestamp_now(),
-                });
-                changed_keys.insert(*key);
-            }
-        }
-
-        if changed_keys.is_empty() {
-            return;
-        }
-
-        self.metrics
-            .set_loc_rib_prefixes("evpn", gauge_val(self.loc_rib.evpn_len()));
-
-        let peers: Vec<IpAddr> = self.outbound_peers.keys().copied().collect();
-        for peer in peers {
-            let sendable = self.peer_sendable_families.get(&peer).cloned();
-            if !sendable.as_ref().is_some_and(|f| {
-                f.contains(&(rustbgpd_wire::Afi::L2Vpn, rustbgpd_wire::Safi::Evpn))
-            }) {
-                continue;
-            }
-
-            let target_is_ebgp = self.peer_is_ebgp.get(&peer).copied().unwrap_or(true);
-            let target_is_rr_client = self.peer_is_rr_client.get(&peer).copied().unwrap_or(false);
-            let target_peer_asn = self.peer_asn.get(&peer).copied();
-            let target_peer_group = self.peer_group.get(&peer).map(String::as_str);
-            let export_pol = self.export_policy_for(peer).cloned();
-            let target_peer_label = peer.to_string();
-            let metrics = self.metrics.clone();
-
-            let loc_rib_len = self.loc_rib.len();
-            let rib_out = self
-                .adj_ribs_out
-                .entry(peer)
-                .or_insert_with(|| crate::adj_rib_out::AdjRibOut::with_capacity(peer, loc_rib_len));
-            let policy_stats = self.export_policy_stats.entry(peer).or_default();
-
-            let mut evpn_announce = Vec::new();
-            let mut evpn_withdraw = Vec::new();
-            Self::stage_evpn_routes(
-                &self.loc_rib,
-                rib_out,
-                &self.peer_is_rr_client,
-                &changed_keys,
-                peer,
-                target_peer_asn,
-                target_peer_group,
-                target_is_ebgp,
-                target_is_rr_client,
-                self.cluster_id,
-                sendable.as_ref(),
-                export_pol.as_ref(),
-                &metrics,
-                policy_stats,
-                &target_peer_label,
-                &mut evpn_announce,
-                &mut evpn_withdraw,
-                false, // EVPN delta path — equality check is correct
-            );
-
-            if (!evpn_announce.is_empty() || !evpn_withdraw.is_empty())
-                && !self.try_send_and_commit_outbound_update(
-                    peer,
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    evpn_announce,
-                    evpn_withdraw,
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                )
-            {
-                warn!(%peer, "outbound channel full — EVPN update deferred");
-                self.dirty_peers.insert(peer);
-            }
-        }
-    }
-
-    /// Recompute Loc-RIB best path and distribute changes for BGP-LS routes.
-    ///
-    /// BGP-LS remains opaque: the selected route is reflected as received with
-    /// ordinary BGP attribute handling, and no LSDB/TLV semantics are parsed.
-    pub(super) fn recompute_and_distribute_bgpls(
-        &mut self,
-        affected: &HashSet<crate::route::BgpLsRouteKey>,
-    ) {
-        use crate::route::BgpLsRibRoute;
-
-        let mut changed_keys: HashSet<crate::route::BgpLsRouteKey> = HashSet::new();
-        for key in affected {
-            let candidates: Vec<BgpLsRibRoute> = self
-                .ribs
-                .values()
-                .filter_map(|rib| rib.get_bgpls(key).cloned())
-                .collect();
-            if self.loc_rib.recompute_bgpls(key.clone(), candidates.iter()) {
-                changed_keys.insert(key.clone());
-            }
-        }
-
-        if changed_keys.is_empty() {
-            return;
-        }
-
-        self.metrics
-            .set_loc_rib_prefixes("bgpls", gauge_val(self.loc_rib.bgpls_len()));
-
-        let peers: Vec<IpAddr> = self.outbound_peers.keys().copied().collect();
-        for peer in peers {
-            let sendable = self.peer_sendable_families.get(&peer).cloned();
-            if !changed_keys.iter().any(|key| {
-                sendable
-                    .as_ref()
-                    .is_some_and(|families| families.contains(&key.family.to_afi_safi()))
-            }) {
-                continue;
-            }
-
-            let target_is_ebgp = self.peer_is_ebgp.get(&peer).copied().unwrap_or(true);
-            let target_is_rr_client = self.peer_is_rr_client.get(&peer).copied().unwrap_or(false);
-            let target_peer_asn = self.peer_asn.get(&peer).copied();
-            let target_peer_group = self.peer_group.get(&peer).map(String::as_str);
-            let export_pol = self.export_policy_for(peer).cloned();
-            let target_peer_label = peer.to_string();
-            let metrics = self.metrics.clone();
-
-            let loc_rib_len = self.loc_rib.len();
-            let rib_out = self
-                .adj_ribs_out
-                .entry(peer)
-                .or_insert_with(|| crate::adj_rib_out::AdjRibOut::with_capacity(peer, loc_rib_len));
-            let policy_stats = self.export_policy_stats.entry(peer).or_default();
-
-            let mut bgpls_announce = Vec::new();
-            let mut bgpls_withdraw = Vec::new();
-            Self::stage_bgpls_routes(
-                &self.loc_rib,
-                rib_out,
-                &self.peer_is_rr_client,
-                &changed_keys,
-                peer,
-                target_peer_asn,
-                target_peer_group,
-                target_is_ebgp,
-                target_is_rr_client,
-                self.cluster_id,
-                sendable.as_ref(),
-                export_pol.as_ref(),
-                &metrics,
-                policy_stats,
-                &target_peer_label,
-                &mut bgpls_announce,
-                &mut bgpls_withdraw,
-                false,
-            );
-
-            if (!bgpls_announce.is_empty() || !bgpls_withdraw.is_empty())
-                && !self.try_send_and_commit_outbound_update(
-                    peer,
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    bgpls_announce,
-                    bgpls_withdraw,
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                )
-            {
-                warn!(%peer, "outbound channel full — BGP-LS update deferred");
-                self.dirty_peers.insert(peer);
-            }
-        }
-    }
-
-    /// Recompute Loc-RIB best path and distribute changes for VPNv4/VPNv6
-    /// routes.
-    ///
-    /// The selected route is reflected as received: Route Distinguisher,
-    /// MPLS label stack, and next-hop pass through unchanged (ADR-0077 §6),
-    /// with ordinary BGP attribute handling applied by transport.
-    pub(super) fn recompute_and_distribute_vpn(
-        &mut self,
-        affected: &HashSet<crate::route::VpnRibRouteKey>,
-    ) {
-        use crate::route::VpnRibRoute;
-
-        let mut changed_keys: HashSet<crate::route::VpnRibRouteKey> = HashSet::new();
-        for key in affected {
-            let candidates: Vec<VpnRibRoute> = self
-                .ribs
-                .values()
-                .filter_map(|rib| rib.get_vpn(key).cloned())
-                .collect();
-            if self.loc_rib.recompute_vpn(key.clone(), candidates.iter()) {
-                changed_keys.insert(key.clone());
-            }
-        }
-
-        if changed_keys.is_empty() {
-            return;
-        }
-
-        self.metrics
-            .set_loc_rib_prefixes("vpn", gauge_val(self.loc_rib.vpn_len()));
-
-        let peers: Vec<IpAddr> = self.outbound_peers.keys().copied().collect();
-        for peer in peers {
-            let sendable = self.peer_sendable_families.get(&peer).cloned();
-            if !changed_keys.iter().any(|key| {
-                sendable
-                    .as_ref()
-                    .is_some_and(|families| families.contains(&key.afi_safi()))
-            }) {
-                continue;
-            }
-
-            let target_is_ebgp = self.peer_is_ebgp.get(&peer).copied().unwrap_or(true);
-            let target_is_rr_client = self.peer_is_rr_client.get(&peer).copied().unwrap_or(false);
-            let target_peer_asn = self.peer_asn.get(&peer).copied();
-            let target_peer_group = self.peer_group.get(&peer).map(String::as_str);
-            let export_pol = self.export_policy_for(peer).cloned();
-            let rtc_filter = self.rtc_vpn_filter(peer, sendable.as_ref());
-            let target_peer_label = peer.to_string();
-            let metrics = self.metrics.clone();
-
-            let loc_rib_len = self.loc_rib.len();
-            let rib_out = self
-                .adj_ribs_out
-                .entry(peer)
-                .or_insert_with(|| crate::adj_rib_out::AdjRibOut::with_capacity(peer, loc_rib_len));
-            let policy_stats = self.export_policy_stats.entry(peer).or_default();
-
-            let mut vpn_announce = Vec::new();
-            let mut vpn_withdraw = Vec::new();
-            Self::stage_vpn_routes(
-                &self.loc_rib,
-                rib_out,
-                &self.peer_is_rr_client,
-                &changed_keys,
-                peer,
-                target_peer_asn,
-                target_peer_group,
-                target_is_ebgp,
-                target_is_rr_client,
-                self.cluster_id,
-                sendable.as_ref(),
-                rtc_filter.as_ref(),
-                export_pol.as_ref(),
-                &metrics,
-                policy_stats,
-                &target_peer_label,
-                &mut vpn_announce,
-                &mut vpn_withdraw,
-                false,
-            );
-
-            if (!vpn_announce.is_empty() || !vpn_withdraw.is_empty())
-                && !self.try_send_and_commit_outbound_update(
-                    peer,
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vpn_announce,
-                    vpn_withdraw,
-                    vec![],
-                    vec![],
-                )
-            {
-                warn!(%peer, "outbound channel full — VPN update deferred");
-                self.dirty_peers.insert(peer);
-            }
-        }
-    }
-
-    /// Recompute Loc-RIB best path and distribute changes for RT-Constrain
-    /// routes (RFC 4684 §3.2). Reflection semantics mirror VPN: the NLRI and
-    /// stored next-hop pass through unchanged.
-    pub(super) fn recompute_and_distribute_rtc(
-        &mut self,
-        affected: &HashSet<crate::route::RtcRibRouteKey>,
-    ) {
-        use crate::route::RtcRibRoute;
-
-        let mut changed_keys: HashSet<crate::route::RtcRibRouteKey> = HashSet::new();
-        for key in affected {
-            let candidates: Vec<RtcRibRoute> = self
-                .ribs
-                .values()
-                .filter_map(|rib| rib.get_rtc(key).cloned())
-                .collect();
-            if self.loc_rib.recompute_rtc(key.clone(), candidates.iter()) {
-                changed_keys.insert(key.clone());
-            }
-        }
-
-        if changed_keys.is_empty() {
-            return;
-        }
-
-        self.metrics
-            .set_loc_rib_prefixes("rtc", gauge_val(self.loc_rib.rtc_len()));
-
-        let rtc_family = crate::route::RtcRibRouteKey::afi_safi();
-        let peers: Vec<IpAddr> = self.outbound_peers.keys().copied().collect();
-        for peer in peers {
-            let sendable = self.peer_sendable_families.get(&peer).cloned();
-            if !sendable
-                .as_ref()
-                .is_some_and(|families| families.contains(&rtc_family))
-            {
-                continue;
-            }
-
-            let target_is_ebgp = self.peer_is_ebgp.get(&peer).copied().unwrap_or(true);
-            let target_is_rr_client = self.peer_is_rr_client.get(&peer).copied().unwrap_or(false);
-            let target_peer_asn = self.peer_asn.get(&peer).copied();
-            let target_peer_group = self.peer_group.get(&peer).map(String::as_str);
-            let export_pol = self.export_policy_for(peer).cloned();
-            let target_peer_label = peer.to_string();
-            let metrics = self.metrics.clone();
-
-            let loc_rib_len = self.loc_rib.len();
-            let rib_out = self
-                .adj_ribs_out
-                .entry(peer)
-                .or_insert_with(|| crate::adj_rib_out::AdjRibOut::with_capacity(peer, loc_rib_len));
-            let policy_stats = self.export_policy_stats.entry(peer).or_default();
-
-            let mut rtc_announce = Vec::new();
-            let mut rtc_withdraw = Vec::new();
-            Self::stage_rtc_routes(
-                &self.loc_rib,
-                rib_out,
-                &self.peer_is_rr_client,
-                &changed_keys,
-                peer,
-                target_peer_asn,
-                target_peer_group,
-                target_is_ebgp,
-                target_is_rr_client,
-                self.cluster_id,
-                sendable.as_ref(),
-                export_pol.as_ref(),
-                &metrics,
-                policy_stats,
-                &target_peer_label,
-                &mut rtc_announce,
-                &mut rtc_withdraw,
-                false,
-            );
-
-            if (!rtc_announce.is_empty() || !rtc_withdraw.is_empty())
-                && !self.try_send_and_commit_outbound_update(
-                    peer,
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    vec![],
-                    rtc_announce,
-                    rtc_withdraw,
-                )
-            {
-                warn!(%peer, "outbound channel full — RTC update deferred");
-                self.dirty_peers.insert(peer);
-            }
-        }
-    }
-}
